JuChrom.jl

Documentation for JuChrom.jl

JuChrom.AbstractChromatogram — Type

AbstractChromatogram

The AbstractChromatogram type is the supertype of all chromatogram implementations in JuChrom. All subtypes of AbstractChromatogram (e.g., FID, GCMS, TIC) contain scantimes, intensities, and source information.

See also AbstractGCMS.

JuChrom.AbstractGCMS — Type

AbstractGCMS <: AbstractChromatogram

The AbstractGCMS type is the supertype of all chromatogram implementations that contain mass-charge ratio (m/z) data (= ions) and associated abundance values (= intensities) and thus can have one or more ion intensity values associated with a given scan time in JuChrom (e.g., GCMS).

See also AbstractChromatogram, GCMS.

JuChrom.GCMS — Method

GCMS(scantimes::AbstractVector{<:Unitful.Time}, ions::AbstractVector{<:Real}, 
intensities::AbstractMatrix{<:Real}; source::Union{Nothing, AbstractString})

Return a GCMS object consisting of scantimes, ions, intensities, and optionally data source information.

See also AbstractGCMS, intensities, ions, scantimes.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the GCMS object preserves the types.

Example

julia> GCMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1])
GCMS {scantimes: Int32, ions: Int64, intensities: Int32}
Source: nothing
3 scans; time range: 1 s - 3 s
2 ions; range: m/z 85 - 100
intensity range: 0 - 956

julia> GCMS([1.1f0, 2.1f0]u"s", Float32[35.1, 76.2], Int64[0 12; 34 956], "example data")
GCMS {scantimes: Float32, ions: Float32, intensities: Int64}
Source: example data
2 scans; time range: 1.1f0 s - 2.1f0 s
2 ions; range: m/z 35.1 - 76.2
intensity range: 0 - 956

JuChrom.intensities — Method

intensities(chrom::AbstractChromatogram)

Return the intensities of an AbstractChromatogram subtype object.

See also AbstractChromatogram.

Example

julia> gcms = GCMS([1u"s", 2u"s", 3u"s"], [85, 100], Int64[0 12; 34 956; 23 1]);

julia> intensities(gcms)
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

JuChrom.ions — Method

ions(gcms::AbstractGCMS)

Return the ions of an AbstractGCMS subtype object (e.g., GCMS`).

See also AbstractGCMS.

In the following example, the number type of ions passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> ions(gcms)
2-element Vector{Int64}:
  85
 100

JuChrom.scantimes — Method

scantimes(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the scantimes of an AbstractChromatogram subtype object (e.g., FID, GCMS, TIC). The optional keyword argument timeunit allows you to change the unit of the returned scantimes. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned values.

See also AbstractChromatogram.

In the following example, the element type of the scantime vector passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS(Float32[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantimes(gcms)
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0f0 s
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, timeunit=u"minute")
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.016666668f0 minute
 0.033333335f0 minute
 0.050000004f0 minute

julia> scantimes(gcms, timeunit=u"minute", ustripped=true)
3-element Vector{Float32}:
 0.016666668
 0.033333335
 0.050000004

JuChrom.source — Method

source(chrom::AbstractChromatogram) -> Type{<:Union{AbstractString, Nothing}}

Return the source information of the AbstractChromatogram subtype object.

See also AbstractChromatogram.

Example

julia> gcms₁ = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1], "example data");

julia> source(gcms₁)
"example data"

julia> gcms₂ = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> source(gcms₂)

julia> isnothing(source(gcms₂))
true