Basics

JuChrom.AbstractChromatogram — Type

AbstractChromatogram

Supertype for all chromatogram implementations. All subtypes (e.g., FID, GCMS, TIC) include scan times, intensities, and metadata.

See also AbstractGC, AbstractFID, AbstractGCMS, AbstractTIC, FID, GCMS, TIC, scantimes, intensities, metadata.

JuChrom.AbstractGC — Type

AbstractGC <: AbstractChromatogram

Supertype for all chromatogram implementations that lack mass-charge ratio (m/z) data (ions) and thus have a single intensity value associated with each scan time (e.g., FID, TIC). The intensities are stored in a vector, with the index corresponding to the scan time index.

See also AbstractChromatogram, AbstractFID, AbstractTIC, FID, TIC, scantimes, intensities, metadata.

JuChrom.AbstractFID — Type

AbstractFID <: AbstractGC

Supertype for all flame ionization detector chromatogram implementations (e.g., FID).

See also AbstractChromatogram, AbstractGC, AbstractFID, FID, scantimes, intensities, metadata.

JuChrom.FID — Type

FID(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any})

Construct a FID object that includes scantimes, intensities, and metadata. Note that scantimes must be in ascending order, and intensities must not contain any values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractFID, RiFID, scantimes, intensities, metadata.

In the following examples, the types of scantimes and intensities are explicitely annotated to demonstrate that the FID object preserves these types.

Examples

julia> FID(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
FID {scan times: Int64, intensities: Int32}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 1 - 956
metadata: 0 entries

julia> FID(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict("name" => "sample"))
FID {scan times: Int32, intensities: Float64}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 1.0 - 956.0
metadata: 1 entry

julia> FID([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scan times not in ascending order
[...]

julia> FID([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.RiFID — Type

RiFID(scantimes::AbstractVector{<:Unitful.Time}, retentionindexname::AbstractString, 
retentionindices::{<:Union{Real, Missing}}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any})

Create an RiFID object that includes scantimes, a retentionindexname, retentionindices, intensities, and metadata. The retentionindices may contain missing values but must have at least one numerical entry. Ensure that scantimes and retentionindices are both in ascending order and have the same length. Additionally, ensure that all values in intensities are non-negative.

See also AbstractChromatogram, AbstractGC, AbstractFID, scantimes, retentionindexname, retentionindices, intensities, metadata.

Examples

julia> RiFID([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> RiFID([1, 2, 3]u"s", "Kovats", [missing, 200.0, 300.0], [12, 956, 1]);

julia> RiFID([1, 2, 3]u"s", "Kovats", [missing, 300.0, 200.0], [12, 956, 1])
ERROR: ArgumentError: retention indices not in ascending order
[...]

julia> RiFID([1, 2, 3]u"s", "Kovats", [100.0, missing, 300.0], [12, 956, -1])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.AbstractTIC — Type

AbstractTIC <: AbstractGC

Supertype for all total ion chromatogram implementations (e.g., TIC).

See also AbstractChromatogram, AbstractGC, AbstractTIC, TIC, scantimes, intensities, metadata.

JuChrom.TIC — Type

TIC(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any}())

Construct a TIC object that includes scantimes, intensities, and metadata. Note that scantimes must be in ascending order, and intensities must not contain any values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractTIC, scantimes, intensities, metadata, totalionchromatogram.

In the following examples, the types of scantimes and intensities are explicitely annotated to demonstrate that the TIC object preserves these types.

Examples

julia> TIC(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
TIC {scan times: Int64, intensities: Int32}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 1 - 956
metadata: 0 entries

julia> TIC(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict("name" => "sample"))
TIC {scan times: Int32, intensities: Float64}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 1.0 - 956.0
metadata: 1 entry

julia> TIC([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scan times not in ascending order
[...]

julia> TIC([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.RiTIC — Type

RiTIC(scantimes::AbstractVector{<:Unitful.Time}, retentionindexname::AbstractString, 
retentionindices::{<:Union{Real, Missing}}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any})

Create an RiTIC object that includes scantimes, a retentionindexname, retentionindices, intensities, and metadata. The retentionindices may contain missing values but must have at least one numerical entry. Ensure that scantimes and retentionindices are both in ascending order and have the same length. Additionally, ensure that all values in intensities are non-negative.

See also AbstractChromatogram, AbstractGC, AbstractTIC, scantimes, retentionindexname, retentionindices, intensities, metadata.

Examples

julia> RiTIC([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> RiTIC([1, 2, 3]u"s", "Kovats", [missing, 200.0, 300.0], [12, 956, 1]);

julia> RiTIC([1, 2, 3]u"s", "Kovats", [missing, 300.0, 200.0], [12, 956, 1])
ERROR: ArgumentError: retention indices not in ascending order
[...]

julia> RiTIC([1, 2, 3]u"s", "Kovats", [100.0, missing, 300.0], [12, 956, -1])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.AbstractGCMS — Type

AbstractGCMS <: AbstractChromatogram

Supertype for all chromatogram implementations that include mass-charge ratio (m/z) data (ions) and associated abundance values (intensities) (e.g., GCMS). This type can have one or more ion intensity values associated with a given scan time. Intensities are stored in a matrix where the row index represents the scan time and the column index represents the ion.

See also AbstractChromatogram, AbstractGCMS, GCMS, scantimes, ions, intensities, metadata.

JuChrom.GCMS — Type

GCMS(scantimes::AbstractVector{<:Unitful.Time}, ions::AbstractVector{<:Real}, 
intensities::AbstractMatrix{<:Real}; metadata::Dict=Dict{Any, Any}())

Construct a GCMS object that includes scantimes, ions, intensities, and metadata. Note that scantimes and ions must be in ascending order, and intensities must not contain any values less than zero.

See also AbstractChromatogram, AbstractGCMS, scantimes, ions, intensities, metadata.

In the following examples, the types of scantimes, ions, and intensities are explicitely annotated to demonstrate that the GCMS object preserves these types.

Examples

julia> GCMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1])
GCMS {scan times: Int32, ions: Int64, intensities: Int32}
3 scans; scan times: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> GCMS([1.1f0, 2.1f0]u"s", [35.1f0, 76.2f0], Int64[0 12; 34 956], Dict(:id => 4))
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2 scans; scan times: 1.1f0 s, 2.1f0 s
2 ions: m/z 35.1, 76.2
intensity range: 0 - 956
metadata: 1 entry

julia> GCMS([2, 1, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
ERROR: ArgumentError: scan times not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [100, 85], [0 12; 34 956; 23 1])
ERROR: ArgumentError: ions not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [85, 100], [0 -12; 34 956; 23 1])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.IonScanOrder — Type

IonScanOrder

Supertype of all ion scan order implementations.

See also LinearAscending, LinearDescending, ionscantimeshift, ionscantimes, ionscantime, ionscantimeindex.

JuChrom.LinearAscending — Type

LinearAscending(; start::Real=0, stop::Real=1) <: IonScanOrder

Construct a LinearAscending ion scan order object. This object specifies that ions were scanned in linear ascending order (i.e., smallest ion first, largest ion last) during each scan. The time allocated to scan each ion is assumed to be equal, resulting from dividing the total scan interval time equally among the ions. The optional start and stop parameters allow you to limit the time interval during which ions were scanned in each scan. These parameters specify relative points within the scan interval: 0 ≤ start < stop ≤ 1. The default values are start=0 and stop=1, meaning the scan of the smallest ion began at the start of the scan interval and the scan of the largest ion concluded at the end of the interval. In contrast, setting the start value to 0.5 indicates that ions were scanned only during the second half of the scan interval. For instance, this could occur if the instrument switched between SIM mode and Scan mode during each scan interval, operating in Scan mode only during the second half, which generated the data in question.

See also AbstractGCMS, LinearDescending, ionscantimeshift, ionscantimes, ionscantime, ionscantimeindex, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Examples

julia> LinearAscending()
LinearAscending{Int64, Int64}(0, 1)

julia> LinearAscending(; start=0.5)
LinearAscending{Float64, Int64}(0.5, 1)

julia> LinearAscending(; start=0.1, stop=0.5)
LinearAscending{Float64, Float64}(0.1, 0.5)

julia> LinearAscending(; start=0.5, stop=0.5)
ERROR: ArgumentError: start=0.5 and stop=0.5 do not satisfy condition 0 ≤ start < stop ≤ 1
[...]

JuChrom.LinearDescending — Type

LinearDescending(; start::Real=0, stop::Real=1) <: IonScanOrder

Construct a LinearDescending ion scan order object. This object specifies that ions were scanned in linear descending order (i.e., largest ion first, smallest ion last) during each scan. The time allocated to scan each ion is assumed to be equal, resulting from dividing the total scan interval time equally among the ions. The optional start and stop parameters allow you to limit the time interval during which ions were scanned in each scan. These parameters specify relative points within the scan interval: 0 ≤ start < stop ≤ 1. The default values are start=0 and stop=1, meaning the scan of the largest ion began at the start of the scan interval and the scan of the smallest ion concluded at the end of the interval. In contrast, setting the start value to 0.5 indicates that ions were scanned only during the second half of the scan interval. For instance, this could occur if the instrument switched between SIM mode and Scan mode during each scan interval, operating in Scan mode only during the second half, which generated the data in question.

See also AbstractGCMS, LinearAscending, ionscantimeshift, ionscantimes, ionscantime, ionscantimeindex, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Examples

julia> LinearDescending()
LinearDescending{Int64, Int64}(0, 1)

julia> LinearDescending(; start=0.5)
LinearDescending{Float64, Int64}(0.5, 1)

julia> LinearDescending(; start=0.1, stop=0.5)
LinearDescending{Float64, Float64}(0.1, 0.5)

julia> LinearDescending(; start=0.5, stop=0.5)
ERROR: ArgumentError: start=0.5 and stop=0.5 do not satisfy condition 0 ≤ start < stop ≤ 1
[...]

JuChrom.binions — Function

binions(gcms::AbstractGCMS; ionbin::Function=integer)

Return a GCMS object in which the ions are binned according to the ionbin function (the default is the integer function), and the intensities of the binned ions are summed.

See also AbstractGCMS, integer, intensities, ions, ioncount.

Examples

julia> gcms = GCMS((1:3)u"s", [84.8, 85.2, 100.9], [0 24 12; 0 0 956; 23 0 1])
GCMS {scan times: Int64, ions: Float64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
3 ions: m/z 84.8, 85.2, 100.9
intensity range: 0 - 956
metadata: 0 entries

julia> intensities(gcms)
3×3 Matrix{Int64}:
  0  24   12
  0   0  956
 23   0    1

julia> gcms_binnedions = binions(gcms)
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
2 ions: m/z 85, 101
intensity range: 0 - 956
metadata: 0 entries

julia> ions(gcms_binnedions)
2-element Vector{Int64}:
  85
 101

julia> intensities(gcms_binnedions)
3×2 Matrix{Int64}:
 24   12
  0  956
 23    1

julia> custom_ionbin(ion) = integer(ion, start=0.9);

julia> gcms_binnedions = binions(gcms, ionbin=custom_ionbin)
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
3 ions: m/z 84, 85, 101
intensity range: 0 - 956
metadata: 0 entries

julia> ions(gcms_binnedions)
3-element Vector{Int64}:
  84
  85
 101

JuChrom.cosine — Function

cosine(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Return the angle between two non-zero vectors, which can be considered a measure of the similarity (i.e., cosine similarity) between the two vectors.

Examples

julia> cosine([100, 500, 250], [200, 1000, 0])
0.8978872704229618

julia> cosine([100, 0, 50], [0, 20, 0])
0.0

julia> cosine([10, 50, 25], [100, 500, 250])
1.0

JuChrom.integer — Function

integer(value:::Real; start::Real=0.7) -> Int

Return the integer for the given value that satisfies the following condition: integer - 1 + start ≤ value < integer + start, where 0 ≤ start < 1.

See also AbstractGCMS, GCMS, binions, ions.

Examples

julia> integer(29.7)
30

julia> integer(30.0)
30

julia> integer(30.69)
30

julia> integer(29.7, start=0.8)
29

JuChrom.intensities — Method

intensities(chrom::AbstractGC; scanindexrange::OrdinalRange{T, S}) where {T<:Integer, 
S<:Integer}

Return the intensities. The optional keyword argument scanindexrange lets you select a subset of scans for which the intensities will be returned. Note that the function will return either a reference to the intensity vector or a view into the intensity vector, depending on whether a subset of scans is selected.

See also AbstractGC, intensity, minintensity, maxintensity, scancount.

Examples

julia> tic = TIC([1, 2, 3]u"s", [123, 224, 103])
TIC {scan times: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensities(tic)  # reference to the data structure
3-element Vector{Int64}:
 123
 224
 103

julia> intensities(tic)[:]  # a copy of these values
3-element Vector{Int64}:
 123
 224
 103

julia> intensities(tic; scanindexrange=2:3)  # view into the data structure
2-element view(::Vector{Int64}, 2:3) with eltype Int64:
 224
 103

julia> intensities(tic; scanindexrange=2:3)[:]  # a copy of these values
2-element Vector{Int64}:
 224
 103

JuChrom.intensities — Method

intensities(gcms::AbstractGCMS; scanindexrange::OrdinalRange{T1, S1}, 
ionindexrange::OrdinalRange{T2, S2}) where {T1<:Integer, S1<:Integer, T2<:Integer, 
S2<:Integer}

Return the intensities. The optional keyword arguments scanindexrange and ionindexrange allow you to select specific parts of the intensity matrix to be returned. Note that the function returns either a reference to the matrix or a view into it, depending on whether the keyword arguments specify subranges of the matrix.

See also AbstractGCMS, scancount, ioncount, minintensity, maxintensity.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensities(gcms)  # reference to the data structure
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

julia> intensities(gcms)[:, :]  # a copy of these values
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

julia> intensities(gcms, ionindexrange=1:1)  # all intensities of the ion at index 1
3×1 view(::Matrix{Int64}, 1:3, 1:1) with eltype Int64:
  0
 34
 23

julia> intensities(gcms, ionindexrange=1:1)[:]  # a copy of these values
3-element Vector{Int64}:
  0
 34
 23

julia> intensities(gcms, scanindexrange=1:1)  # intensities of all ions from scan 1
1×2 view(::Matrix{Int64}, 1:1, 1:2) with eltype Int64:
 0  12

julia> intensities(gcms, scanindexrange=1:1)[:]  # a copy of these values
2-element Vector{Int64}:
  0
 12

julia> intensities(gcms, scanindexrange=1:2, ionindexrange=1:2)
2×2 view(::Matrix{Int64}, 1:2, 1:2) with eltype Int64:
  0   12
 34  956

julia> intensities(gcms, scanindexrange=1:2, ionindexrange=1:2)[:, :]
2×2 Matrix{Int64}:
  0   12
 34  956

JuChrom.intensity — Method

intensity(chrom::AbstractGC, scanindex::Integer)

Return the intensity for a scan by specifying its scanindex.

See also AbstractGC, intensities, minintensity, maxintensity.

Examples

julia> tic = TIC([1, 2, 3]u"s", [123, 224, 103])
TIC {scan times: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensity(tic, 1)
123

julia> intensity(tic, 2)
224

JuChrom.intensity — Method

intensity(chrom::AbstractGC, time::Unitful.Time; precisetime::Bool=false)

Return the intensity at a given time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. By default, the intensity associated with the scan whose timestamp is closest to the given time is returned. In case of a tie, the intensity of the scan with the later timestamp is used. If the optional parameter precisetime is set to true, the specified time must exactly match a timestamp in the vector; otherwise, an error is thrown.

See also AbstractGC, intensities, scantimes, minscantime, maxscantime.

Examples

julia> tic = TIC([1.0, 2.0, 3.0]u"s", [123, 224, 103])
TIC {scan times: Float64, intensities: Int64}
3 scans; scan times: 1.0 s, 2.0 s, 3.0 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensity(tic, 1.5u"s")
224

julia> intensity(tic, 1u"s", precisetime=true)
123

julia> intensity(tic, 1.5u"s", precisetime=true)
ERROR: ArgumentError: scantime 1.5 s does not exist
[...]

JuChrom.intensity — Method

intensity(gcms::AbstractGCMS, scanindex::Integer, ionindex::Integer)

Return the intensity of an ion in a scan, given the scanindex of the scan and the ionindex of the ion.

See also AbstractGCMS, scancount, ions, ioncount.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensity(gcms, 2, 1)
34

julia> intensity(gcms, 1, 2)
12

JuChrom.intensity — Method

intensity(gcms::AbstractGCMS, time::Unitful.Time, ion::Real; precisetime::Bool=false)

Return the intensity of an ion at a given time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. By default, the intensity associated with the scan whose timestamp is closest to the given time is returned. In case of a tie, the intensity of the scan with the later timestamp is used. If the optional parameter precisetime is set to true, the specified time must exactly match a timestamp in the vector; otherwise, an error is thrown.

See also AbstractGCMS, intensities, ions, scantimes.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Float64, ions: Int64, intensities: Int64}
3 scans; scan times: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensity(gcms, 1.9u"s", 85)
34

julia> intensity(gcms, 2.9u"s", 85, precisetime=true)
ERROR: ArgumentError: scantime 2.9 s does not exist
[...]

julia> intensity(gcms, 3u"s", 85, precisetime=true)
23

JuChrom.ion — Function

ion(gcms::AbstractGCMS, ionindex::Integer)

Return the ion at the specified ionindex.

See also AbstractGCMS, ions, ionindex, minion, maxion, ioncount.

Examples

julia> gcms = GCMS((1:3)u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> ion(gcms, 1)
85

julia> ion(gcms, 2)
100

julia> ion(gcms, 3)
ERROR: BoundsError: attempt to access 2-element Vector{Int64} at index [3]
[...]

JuChrom.ioncount — Function

ioncount(gcms::AbstractGCMS) -> Int

Return the number of ions.

See also AbstractGCMS, ions, ion, minion, maxion.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> ioncount(gcms)
2

JuChrom.ionindex — Function

ionindex(gcms::AbstractGCMS, ion::Real) -> Int

Return the index of the specified ion. An error is thrown if the ion does not exist.

See also AbstractGCMS, ions, ioncount, ion, minion, maxion, ionscantime, ionscantimeshift.

Examples

julia> gcms = GCMS((1:3)u"s", [85.2f0, 100.1f0], [0 12; 34 956; 23 1]);

julia> ionindex(gcms, 100.1)
2

julia> ionindex(gcms, 201.1)
ERROR: ArgumentError: ion 201.1 does not exist
[...]

JuChrom.ionscantime — Function

ionscantime(δtᵢ::Function, gcms::AbstractGCMS, scanindex::Integer, ionindex::Integer; 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the time at which an ion was actually scanned, given the scanindex, ionindex, and a function δtᵢ that computes the time difference between the timestamp of a scan and the scan time of the ion from the ionindex. The optional parameter timeunit allows you to specify the unit of the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether the unit is included in the returned value. Note that the timestamp of a scan is assumed to be the time when the scanning of ion intensities associated with that scan was completed.

See also AbstractGCMS, scantimes, scantime, scantimeindex, ions, ionindex, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Float64, ions: Int64, intensities: Int64}
3 scans; scan times: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantime(δtᵢ, gcms, 2, 1)
2.0 s

julia> ionscantime(δtᵢ, gcms, 2, 2)
1.5 s

julia> ionscantime(δtᵢ, gcms, 2, 2; timeunit=u"minute")
0.025 minute

julia> ionscantime(δtᵢ, gcms, 2, 2; timeunit=u"minute", ustripped=true)
0.025

JuChrom.ionscantimeindex — Function

ionscantimeindex(δtᵢ::Function, gcms::AbstractGCMS, ionindex::Integer, 
time::Unitful.Time; precisetime::Bool=false) -> Int

Return the index of the scan where the scan time for the ion is closest to the specified time, given the ionindex and a function δtᵢ that computes the time difference between the timestamp of a scan and the scan time of the ion from the ionindex. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. In case of a tie, the larger scan index is returned. If the optional parameter precisetime is set to true, the ion must have been scanned exactly at the specified time; otherwise, an error is thrown.

See also AbstractGCMS, scantimeindex, ionscantime, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending, scantimes, scantime, ions, ion, ionindex.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
3 scans; scan times: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantime(δtᵢ, gcms, 2, 2)
1.5 s

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.5u"s")
2

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.6u"s")
2

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.6u"s", precisetime=true)
ERROR: ArgumentError: ion has not been scanned at the time 1.6 s
[...]

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.5u"s", precisetime=true)
2

julia> ionscantimeindex(δtᵢ, gcms, 1, 1.4u"s")
1

julia> ionscantime(δtᵢ, gcms, 1, 1)
1.0 s

JuChrom.ionscantimes — Function

ionscantimes(δtᵢ::Function, gcms::AbstractGCMS, ionindex::Integer; 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the times at which an ion was actually scanned, given the ionindex and a function δtᵢ that computes the time difference between the timestamp of a scan and the scan time of the ion from the ionindex. The optional parameter timeunit allows you to specify the unit for the returned scan times. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether to include the unit in the returned value. Note that the timestamp of a scan is assumed to be the time when the scanning of ion intensities associated with that scan was completed.

See also AbstractGCMS, ionscantime, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending, scantimes, scantimeindex, ions, ionindex.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Float64, ions: Int64, intensities: Int64}
3 scans; scan times: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantimes(δtᵢ, gcms, 1)
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0 s
 2.0 s
 3.0 s

julia> ionscantimes(δtᵢ, gcms, 2)
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 0.5 s
 1.5 s
 2.5 s

julia> ionscantimes(δtᵢ, gcms, 2; timeunit=u"minute")
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.008333333333333333 minute
                0.025 minute
 0.041666666666666664 minute

julia> ionscantimes(δtᵢ, gcms, 2; timeunit=u"minute", ustripped=true)
3-element Vector{Float64}:
 0.008333333333333333
 0.025
 0.041666666666666664

JuChrom.ionscantimeshift — Function

ionscantimeshift(gcms::AbstractGCMS, ionscanorder::IonScanOrder; error::Real=0.001)

Return a function, based on the ionscanorder, that calculates the time difference between the timestamp of a scan and the time when an ion was actually scanned, given the ion index as an argument. The time difference will be zero or negative, since the timestamp of a scan is considered to be when the scanning of the last ion was completed. The returned function assumes that the duration of each scan is consistent throughout the run. The optional keyword argument error lets you specify the maximum allowed deviation of the scan duration, as a fraction of the average scan time, between the timestamps of two consecutive scans.

See also AbstractGCMS, IonScanOrder, LinearAscending, LinearDescending, ionscantime, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Float64, ions: Int64, intensities: Int64}
3 scans; scan times: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearAscending());

julia> δtᵢ(1)
-0.5 s

julia> δtᵢ(2)
0.0 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> δtᵢ(1)
0.0 s

julia> δtᵢ(2)
-0.5 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(start=0.5));

julia> δtᵢ(1)
0.0 s

julia> δtᵢ(2)
-0.25 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(stop=0.5));

julia> δtᵢ(1)
-0.5 s

julia> δtᵢ(2)
-0.75 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(start=0.25, stop=0.75));

julia> δtᵢ(1)
-0.25 s

julia> δtᵢ(2)
-0.5 s

JuChrom.ions — Function

ions(gcms::AbstractGCMS)

Return the ions.

See also AbstractGCMS, ioncount, ion, minion, maxion, ionindex, ionscantime, ionscantimeshift.

In the following examples, the type of ions is explicitely annotated to demonstrate that the GCMS object preserves this type.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> ions(gcms)
2-element Vector{Int64}:
  85
 100

julia> gcms = GCMS([1, 2, 3]u"s", Float32[85, 100], [0 12; 34 956; 23 1]);

julia> ions(gcms)
2-element Vector{Float32}:
  85.0
 100.0

JuChrom.maxintensity — Function

maxintensity(chrom::AbstractChromatogram)

Return the maximum intensity.

See also AbstractChromatogram, minintensity, intensities, intensity.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxintensity(gcms)
956

JuChrom.maxion — Function

maxion(gcms::AbstractGCMS)

Return the largest ion.

See also AbstractGCMS, minion, ions, ion, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxion(gcms)
100

JuChrom.maxretentionindex — Function

maxretentionindex(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{T<:Integer, 
<:Integer}])

Return the maximum retention index, ignoring any missing values. An optional second argument lets you specify a range of scans for which the maximum retention index will be returned. Note that this function only applies to certain AbstractChromatogram subtypes that store retention index data, such as RiFID.

See also AbstractChromatogram, RiFID, minretentionindex, retentionindices, retentionindexname.

Examples

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> maxretentionindex(rifid)
300

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", [100.0, 200.0, missing], [12, 956, 1]);

julia> maxretentionindex(rifid, 2:3)
200.0

JuChrom.maxscantime — Function

maxscantime(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}]; timeunit::Unitful.TimeUnits, ustripped::Bool)

Return the maximum scan time. The optional second positional argument allows you to specify the scan range for which the maximum scan time is returned. The optional keyword argument timeunit lets you change the unit of the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether to include the unit in the returned value.

See also AbstractChromatogram, minscantime, scantimes, scantime,scancount.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxscantime(gcms)
3.0 s

julia> maxscantime(gcms, timeunit=u"minute")
0.05 minute

julia> maxscantime(gcms, timeunit=u"minute", ustripped=true)
0.05

julia> maxscantime(gcms, 1:2)
2.0 s

julia> maxscantime(gcms, 1:2, timeunit=u"minute", ustripped=true)
0.03333333333333333

JuChrom.metadata — Function

metadata(chrom::AbstractChromatogram) -> Dict{Any, Any}

Return the metadata.

See also AbstractChromatogram.

Examples

julia> gcms₁ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956], Dict(:id => 4))
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
2 scans; scan times: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 1 entry

julia> metadata(gcms₁)
Dict{Any, Any} with 1 entry:
  :id => 4

julia> gcms₂ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956])
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
2 scans; scan times: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> metadata(gcms₂)
Dict{Any, Any}()

julia> metadata(gcms₂)["name"] = "sample"
"sample"

julia> metadata(gcms₂)[:id] = 123
123

julia> metadata(gcms₂)
Dict{Any, Any} with 2 entries:
  "name" => "sample"
  :id    => 123

julia> gcms₂
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
2 scans; scan times: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 2 entries

julia> delete!(metadata(gcms₂), "name")
Dict{Any, Any} with 1 entry:
  :id => 123

julia> gcms₂
GCMS {scan times: Int64, ions: Int64, intensities: Int64}
2 scans; scan times: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 1 entry

JuChrom.minintensity — Function

minintensity(chrom::AbstractChromatogram[, greaterthan::Real])

Return the minimum intensity. The optional argument greaterthan allows you to specify a value; the returned minimum intensity will be greater than this value. If no intensity above the specified value is found, nothing is returned.

See also AbstractChromatogram, maxintensity, intensities, intensity.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minintensity(gcms)
0

julia> minintensity(gcms, 0)
1

julia> minintensity(gcms, 1000)

JuChrom.minion — Function

minion(gcms::AbstractGCMS)

Return the smallest ion.

See also AbstractGCMS, maxion, ions, ion, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minion(gcms)
85

JuChrom.minretentionindex — Function

minretentionindex(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{T<:Integer, 
<:Integer}])

Return the minimum retention index, ignoring any missing values. An optional second argument lets you specify a range of scans for which the minimum retention index will be returned. Note that this function only applies to certain AbstractChromatogram subtypes that store retention index data, such as RiFID.

See also AbstractChromatogram, RiFID, maxretentionindex, retentionindices, retentionindexname.

Examples

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> minretentionindex(rifid)
100

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", [100.0, 200.0, missing], [12, 956, 1]);

julia> minretentionindex(rifid, 2:3)
200.0

JuChrom.minscantime — Function

minscantime(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}}]; timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the minimum scan time. The optional second positional argument scanindexrange allows you to specify a scan range for which the minimum scan time is returned. The optional keyword argument timeunit lets you change the unit of the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether to include the unit in the returned value.

See also AbstractChromatogram, maxscantime, scantimes, scantime, scancount.

Examples

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minscantime(gcms)
1.0 s

julia> minscantime(gcms, timeunit=u"minute")
0.016666666666666666 minute

julia> minscantime(gcms, timeunit=u"minute", ustripped=true)
0.016666666666666666

julia> minscantime(gcms, 2:3)
2.0 s

julia> minscantime(gcms, 2:3, timeunit=u"minute", ustripped=true)
0.03333333333333333

JuChrom.retentionindices — Function

retentionindices(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}])

Return the retention indices. An optional second argument lets you specify a range of scans for which the retention indices will be returned. Note that this function only applies to certain AbstractChromatogram subtypes that store retention index data, such as RiFID. Additionally, be aware that the returned retention indices may include missing values. Depending on whether a subset of scans is selected, the function will return either a reference to the data structure holding the retention indices or a view into it.

See also AbstractChromatogram, RiFID, retentionindexname, maxretentionindex, minretentionindex.

Examples

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> retentionindices(rifid)
3-element Vector{Int64}:
 100
 200
 300

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", [missing, 200.0, 300.0], [12, 956, 1]);

julia> retentionindices(rifid)
3-element Vector{Union{Missing, Float64}}:
    missing
 200.0
 300.0

julia> retentionindices(rifid, 2:3)  # view into the data structure
2-element view(::Vector{Union{Missing, Float64}}, 2:3) with eltype Union{Missing, Float64}:
 200.0
 300.0

julia> retentionindices(rifid, 2:3)[:]  # a copy of these values
2-element Vector{Union{Missing, Float64}}:
 200.0
 300.0

JuChrom.retentionindexname — Function

retentionindexname(chrom::AbstractChromatogram)

Return the retention index name. Note that this function only applies to certain AbstractChromatogram subtypes that store retention index data, such as RiFID.

See also AbstractChromatogram, RiFID, retentionindices, maxretentionindex, minretentionindex.

Example

julia> rifid = RiFID([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> retentionindexname(rifid)
"Kovats"

JuChrom.runduration — Function

runduration(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the duration of the run. The optional keyword argument timeunit lets you specify the unit for the returned time interval. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether to include the unit in the returned value.

See also AbstractChromatogram, scantimes, minscantime, maxscantime, scancount.

Examples

julia> fid = FID([30.1u"minute", 40.8u"minute", 51.5u"minute"], [12, 956, 23])
FID {scan times: Float64, intensities: Int64}
3 scans; scan times: 30.1 minute, 40.8 minute, 51.5 minute
intensity range: 12 - 956
metadata: 0 entries

julia> runduration(fid)
21.4 minute

julia> runduration(fid, timeunit=u"s")
1284.0 s

julia> runduration(fid, timeunit=u"s", ustripped=true)
1284.0

JuChrom.scancount — Function

scancount(chrom::AbstractChromatogram) -> Int

Return the number of scans.

See also AbstractChromatogram, scantimes.

Example

julia> fid = FID([1, 2, 3]u"s", [12, 956, 23]);

julia> scancount(fid)
3

JuChrom.scanduration — Function

scanduration(chrom::AbstractChromatogram; error::Real=0.001, 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Calculates the periodicity with which the scans were recorded over time. The optional keyword argument error allows you to specify the maximum allowable deviation of the time interval between consecutive scans from the average scan time, as a fraction of the average scan time. The optional keyword argument timeunit lets you specify the unit for the returned value. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether the unit is include in the returned value.

See also AbstractChromatogram, scantimes, minscantime, maxscantime, scancount, runduration.

Examples

julia> fid = FID([1.0, 2.0, 3.0]u"s", [12, 956, 1]);

julia> scanduration(fid)
1.0 s

julia> scanduration(fid, timeunit=u"minute")
0.016666666666666666 minute

julia> scanduration(fid, timeunit=u"minute", ustripped=true)
0.016666666666666666

julia> scanduration(FID([1.0, 1.99, 3.0]u"s", [12, 956, 1]))
ERROR: ArgumentError: maximum scan duration variation above threshold: 0.010000000000000009 > 0.001
[...]

julia> scanduration(FID([1.0, 1.99, 3.0]u"s", [12, 956, 1]), error=0.02)
1.0 s

JuChrom.scantime — Function

scantime(chrom::AbstractChromatogram, scanindex::Integer; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the scan time for a given scanindex. The optional parameter timeunit lets you specify the unit for the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether the unit is included in the returned value.

See also AbstractChromatogram, scantimes, minscantime, maxscantime, scancount, ionscantime.

Examples

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantime(gcms, 2)
2 s

julia> scantime(gcms, 2, timeunit=u"minute")
1//30 minute

julia> scantime(gcms, 2, timeunit=u"minute", ustripped=true)
1//30

JuChrom.scantimeindex — Function

scantimeindex(gcms::AbstractChromatogram, time::Unitful.Time; 
precisetime::Bool=false) -> Int

Return the index of the scan time closest to time in the scan times. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. In case of a tie, the larger scan time is returned. If the optional parameter precisetime is set to true, the specified time must exactly match a scan time value; otherwise, an error is thrown.

See also AbstractChromatogram, scantimes, scantime, minscantime, maxscantime, scancount, ionindex.

Examples

julia> gcms = GCMS([1.1f0, 2.1f0, 3.1f0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scan times: Float32, ions: Int64, intensities: Int64}
3 scans; scan times: 1.1f0 s, 2.1f0 s, 3.1f0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> scantimeindex(gcms, 1.1f0u"s", precisetime=true)
1

julia> scantimeindex(gcms, 2.1u"s", precisetime=true)
2

julia> scantimeindex(gcms, 2.2u"s", precisetime=true)
ERROR: ArgumentError: scantime 2.2 s does not exist
[...]

julia> scantimeindex(gcms, 2.2u"s")
2

JuChrom.scantimes — Function

scantimes(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}]; timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the scan times. The optional second positional argument allows you to specify a range of scan indices. The optional keyword argument timeunit lets you change the unit of the returned scan times. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped lets you choose whether to include the unit in the returned values. Note: If no time unit conversion is applied and the unit is not stripped, the function returns a reference to or a view into the data structure.

See also AbstractChromatogram, scantime, minscantime, maxscantime, scancount.

In the following example, the type of scantimes is explicitely annotated to demonstrate that the GCMS object preserves this type.

Example

julia> gcms = GCMS(Float32[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantimes(gcms)  # reference to the data structure
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0f0 s
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, timeunit=u"minute")
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.016666668f0 minute
 0.033333335f0 minute
 0.050000004f0 minute

julia> scantimes(gcms, timeunit=u"minute", ustripped=true)
3-element Vector{Float32}:
 0.016666668
 0.033333335
 0.050000004

julia> scantimes(gcms, 2:3)  # view into the data structure
2-element view(::Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}, 2:3) with eltype Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}:
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, 2:3)[:]  # a copy of these values
2-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, 2:3, timeunit=u"minute", ustripped=true)
2-element Vector{Float32}:
 0.033333335
 0.050000004

JuChrom.totalionchromatogram — Function

totalionchromatogram(gcms::GCMS)

Compute the total ion chromatrogram.

See also GCMS, scantimes, intensities, metadata.

In the following example, the type of the intensities is explicitely annotated to demonstrate that the TIC object preserves this type.

Example

julia> gcms = GCMS((1.1f0:3.1f0)u"s", [85, 100], Int64[0 12; 34 956; 23 1])
GCMS {scan times: Float32, ions: Int64, intensities: Int64}
3 scans; scan times: 1.1f0 s, 2.1f0 s, 3.1f0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> tic = totalionchromatogram(gcms)
TIC {scan times: Float32, intensities: Int64}
3 scans; scan times: 1.1f0 s, 2.1f0 s, 3.1f0 s
intensity range: 12 - 990
metadata: 0 entries

julia> intensities(tic)
3-element Vector{Int64}:
  12
 990
  24

julia> gcms = GCMS((1.1:3.1)u"s", [85, 100], Float64[0 12; 34 956; 23 1])
GCMS {scan times: Float64, ions: Int64, intensities: Float64}
3 scans; scan times: 1.1 s, 2.1 s, 3.1 s
2 ions: m/z 85, 100
intensity range: 0.0 - 956.0
metadata: 0 entries

julia> tic = totalionchromatogram(gcms)
TIC {scan times: Float64, intensities: Float64}
3 scans; scan times: 1.1 s, 2.1 s, 3.1 s
intensity range: 12.0 - 990.0
metadata: 0 entries

julia> intensities(tic)
3-element Vector{Float64}:
  12.0
 990.0
  24.0