Data export

JuChrom.InputOutput.DelimitedTextWriter.DelimitedText — Type

DelimitedText(; delim::Union{AbstractChar, AbstractString}="\t") <: Fileformat

Return a DelimitedText file format object. The optional delim keyword argument allows you to specify the column delimiter, which can be either a single character or a string. If the provided filename lacks a suffix, an appropriate one will be automatically appended.

See also FileFormat, exportdata.

Examples

julia> DelimitedText()
DelimitedText{String}("	")

julia> DelimitedText(delim=';')
DelimitedText{Char}(';')

JuChrom.InputOutput.ExcelWriter.Excel — Type

Excel(; sheetname::AbstractString) <: FileFormat

Returns an Excel file format object. The optional keyword argument sheetname allows you to specify the name of the sheet to operate on.

See also FileFormat, exportdata.

Examples

julia> Excel()
Excel{Nothing}(nothing)

julia> Excel(sheetname="TIC")
Excel{String}("TIC")

JuChrom.InputOutput.exportdata — Function

exportdata(chrom::AbstractChrom, file::AbstractString, fileformat::FileFormat; 
timeunit::Unitful.TimeUnits, overwrite::Bool=false)

Write the scan times and associated intensity values of the AbstractChrom object to a file in the specified format. The optional parameter timeunit allows you to specify the unit for the exported scan times. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument overwrite allows to specify whether an existing target file should be overwritten.

See also AbstractChrom, DelimitedText, Excel.

Examples

julia> chrom = Chrom((1:3)u"minute", [123, 224, 103])
Chrom {scan times: Int64, intensities: Int64}
3 scans; scan times: 1 minute, 2 minute, 3 minute
intensity range: 103 - 224
metadata: 0 entries

julia> exportdata(chrom, "./delimtest", DelimitedText());

julia> exportdata(chrom, "./delimtest", Excel(sheetname="TIC"));