Data import

JuChrom.InputOutput.FileFormat — Type

FileFormat

Supertype of all FileFormat implementations.

See also AgilentFID, ChemStationMS.

JuChrom.InputOutput.AgilentFIDReaders.AgilentFID — Type

AgilentFID(; datafilename::AbstractString="FID1A.ch")

Returns an object representing the AgilentFID file format. The optional keyword argument, datafilename, allows you to specify a different name for the data file to be read, instead of the default FID1A.ch. Upper and lower case are not considered when specifying the data file name. Note that when using the function importdata, the AgilentFID data reader expects the Agilent .D folder as the source location, with the specified data file located at the top level of the .D folder. However, you can also provide the full pathname of the data file directly when using the function importdata, in which case the datafilename entry in the AgilentFID object is ignored.

See also FileFormat, importdata.

Examples

julia> AgilentFID()
AgilentFID{String}("FID1A.ch")

julia> AgilentFID(datafilename="fid_data_20240611.ch")
AgilentFID{String}("fid_data_20240611.ch")

julia> dfolder = joinpath(JuChrom.agilent, "ZK_ONUBE_Mix1_11.D");

julia> fid = importdata(dfolder, AgilentFID())
FID {scan times: Float32, intensities: Float64}
4151 scans; scan time range: 0.437f0 ms - 830000.44f0 ms
intensity range: 0.0 - 1.0738316309895832e6
metadata: 10 entries

JuChrom.InputOutput.ChemStationMSReaders.ChemStationMS — Type

ChemStationMS(; datafilename::AbstractString="data.ms")

Returns an object representing the ChemStationMS file format. The optional keyword argument, datafilename, allows you to specify a different name for the data file to be read, instead of the default data.ms (e.g., datasim.ms). Upper and lower case are not considered when specifying the data file name. Note that when using the function importdata, the ChemStationMS data reader expects the Agilent .D folder as the source location, with the specified data file located at the top level of the .D folder. However, you can also provide the full pathname of the data file directly when using the function importdata, in which case the datafilename entry in the ChemStationMS object is ignored.

See also FileFormat, importdata.

Examples

julia> ChemStationMS()
ChemStationMS{String}("data.ms")

julia> ChemStationMS(datafilename="DATASIM.MS")
ChemStationMS{String}("DATASIM.MS")

julia> dfolder = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D");

julia> gcms = importdata(dfolder, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 10 entries

julia> datafile = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D/data.ms");

julia> gcms = importdata(datafile, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 10 entries

JuChrom.InputOutput.MassHunterMSReaders.MassHunterMS — Type

MassHunterMS(; scanmethodid::Integer=1, scantimedigits::Integer=1, 
iondigits::Integer=1, intensitydigits::Integer=0)

Return an object representing the MassHunterMS file format. The optional scanmethodid keyword argument allows you to specify which scan method ID's data to read, especially when multiple scan methods have been applied simultaneously. The optional scantimedigits keyword argument allows you to round the scan time to a specified number of decimal digits. Similarly, the optional iondigits keyword argument rounds the mass-to-charge ratio values to the desired number of decimal places. Note that this is different from ion binning. The optional intensitydigits keyword argument rounds the intensity values to the specified number of decimals. If zero decimal places are specified, the values are returned as integers. The keyword arguments scantimedigits, mzdigits, and intensitydigits address the inconvenience that MassHunter files may store data in types different than the original data collected. A simple type conversion would be insufficient due to inaccuracies introduced during the data conversion before storing the data in MassHunter files. Note that when using the function importdata, the MassHunterMS data reader expects the Agilent .D folder as the source location.

See also FileFormat, importdata, JuChrom.binions.

Examples

julia> dfolder = joinpath(JuChrom.agilent, "C7-C40_MassHunterMS.D");

julia> gcms = importdata(dfolder, MassHunterMS())
GCMS {scan times: Float64, ions: Float64, intensities: Int64}
2405 scans; scan time range: 3.199 minute - 31.651 minute
5174 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1187248
metadata: 0 entries

julia> gcms = importdata(dfolder, MassHunterMS(intensitydigits=1))
GCMS {scan times: Float64, ions: Float64, intensities: Float64}
2405 scans; scan time range: 3.199 minute - 31.651 minute
5174 ions; range: m/z 29.0 - 562.9
intensity range: 0.0 - 1.1872475e6
metadata: 0 entries

JuChrom.InputOutput.importdata — Function

importdata(source::AbstractString, fileformat::FileFormat)

Read data from the specified source using the appropriate file format reader. The source may be a file or a directory path, depending on the requirements of the chosen file format reader.

See also AgilentFID, ChemStationMS.

Example

julia> dfolder = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D");

julia> gcms = importdata(dfolder, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 10 entries