Internals

JuChrom.bsplineinterpolation — Function

bsplineinterpolation(retentiontimes::AbstractVector{<:Unitful.Time}, 
retentionindices::AbstractVector{<:Real}; extrapolation::Bool=true, force::Bool=false)

Return a function that maps retention time to a retention index. The function uses a B-spline for interpolation calculated from a vector of retention times and a corresponding vector of retention indices. For retention time values outside the range used to compute the B-spline, the function employs linear extrapolation to estimate a retention index. However, an optional keyword argument, extrapolation, can be used to disable extrapolation, in which case the function returns nothing for values outside the retention time range. The function will raise an error if the resulting mapping function does not produce continuously increasing values. However, by setting the force keyword argument to true, the function will return a mapping function even if it does not yield continuously increasing values. This can be useful for identifying problematic or erroneous calibration points.

See also scantimes, retentionindices.

Examples

julia> rts = [1, 2, 3, 4, 5, 6, 7, 8]*u"s";

julia> ris = [1000, 1800, 3050, 3800, 5500, 6600, 6900, 7400];

julia> rt2ri = JuChrom.bsplineinterpolation(rts, ris);

julia> rt2ri(1u"s") ≈ 1000.0
true

julia> rt2ri(1.5u"s") ≈ 1333.7469941600825
true

julia> rt2ri((1//30)u"minute") ≈ 1800.0
true

julia> rt2ri(9.1u"s") ≈ 7950.0
true

julia> rt2ri = JuChrom.bsplineinterpolation(rts, ris; extrapolation=false);

julia> rt2ri(9.1u"s") === nothing
ERROR: ArgumentError: retention time is outside the range for calculating the retention index
[...]

JuChrom.piecewiselinearinterpolation — Function

piecewiselinearinterpolation(retentiontimes::AbstractVector{<:Unitful.Time}, 
retentionindices::AbstractVector{<:Real}; extrapolation::Bool=true) -> Float64

Return a function that maps retention time to retention index using piecewise linear interpolation based on a vector of retention times and a corresponding vector of retention indices. For retention time values outside the interpolation range, the function applies linear extrapolation to estimate the retention index. However, an optional extrapolation keyword can disable extrapolation, in which case the function will raise an error for values outside the retention time range.

See also scantimes, retentionindices.

Examples

julia> rts = [1, 2, 3, 4, 5, 6, 7, 8]*u"s";

julia> ris = [1000, 1800, 3050, 3800, 5500, 6600, 6900, 7400];

julia> rt2ri = JuChrom.piecewiselinearinterpolation(rts, ris);

julia> rt2ri(1u"s") ≈ 1000.0
true

julia> rt2ri(1.5u"s") ≈ 1400.0
true

julia> rt2ri((1//30)u"minute") ≈ 1800.0
true

julia> rt2ri(9.1u"s") ≈ 7950.0
true

julia> rt2ri = JuChrom.piecewiselinearinterpolation(rts, ris; extrapolation=false);

julia> rt2ri(9.1u"s")
ERROR: ArgumentError: retention time is outside the range for calculating the retention index
[...]

JuChrom.copy_with_eltype — Function

JuChrom.copy_with_eltype(array::AbstractArray, elementtype::Type)

Create a mutable copy of the array with the type of its elements converted to elementtype.

Examples

julia> JuChrom.copy_with_eltype(Int[1, 2, 3, 4, 5, 6], Float64)
6-element Vector{Float64}:
 1.0
 2.0
 3.0
 4.0
 5.0
 6.0

julia> JuChrom.copy_with_eltype(Float64[1, 2, 3, 4, 5, 6], Int32)
6-element Vector{Int32}:
 1
 2
 3
 4
 5
 6

julia> JuChrom.copy_with_eltype(Float64[1.1, 2, 3, 4, 5, 6], Int32)
ERROR: InexactError: Int32(1.1)
[...]

JuChrom.findclosest — Function

JuChrom.findclosest(A::AbstractVector{<:Number}, x::Number) -> Int

Return the index of the number closest to x in a list A of numbers sorted in ascending order. If case of a tie, the index of the larger number is returned.

Examples

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], 0)
3

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], 1.5)
5

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], -1.5)
2

JuChrom.invert — Function

JuChrom.invert(dictionary::Dict)

Return a dictionary where the values from the input dictionary become the keys. Each key in the returned dictionary maps to a list of all original keys from dictionary that were associated with that value.

Example

julia> d = Dict(:a => 1.0, :b => 2.0, :c => 2.0, :d => 1.0)
Dict{Symbol, Float64} with 4 entries:
  :a => 1.0
  :b => 2.0
  :d => 1.0
  :c => 2.0

julia> JuChrom.invert(d)
Dict{Float64, Vector{Symbol}} with 2 entries:
  2.0 => [:b, :c]
  1.0 => [:a, :d]

JuChrom.metricprefix — Function

JuChrom.metricprefix(number::Real) -> Tuple{String, Int}

Return the metric prefix and the corresponding exponent for a given number.

Examples

julia> JuChrom.metricprefix(347 * 10^7)
("G", 9)

julia> JuChrom.metricprefix(-42558335)
("M", 6)

julia> JuChrom.metricprefix(23)
("", 0)

julia> JuChrom.metricprefix(-0.00003623)
("µ", -6)

julia> JuChrom.metricprefix(425 * 10^-17)
("f", -15)

julia> JuChrom.metricprefix(big"136455347543543453485634875672536725423547234")
("Q", 30)

julia> JuChrom.metricprefix(Inf)
("", 0)

julia> JuChrom.metricprefix(0)
("", 0)

julia> JuChrom.metricprefix(NaN)
("", 0)

JuChrom.name — Function

JuChrom.name(::Type)

Return the name of the type.

Example

julia> chrom = ChromMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1]);

julia> JuChrom.name(typeof(chrom))
ChromMS

JuChrom.rt2ri — Function

JuChrom.rt2ri(mapper::RiMapper)

Return the function that maps a retention time to its corresponding retention index. Note that direct use of this function is discouraged; users are encouraged to use the more versatile function retentionindex instead.

See also RiMapper, retentionindex.

Example

julia> ld = RiMapper("Kovats", [1.2, 2.4, 4.3]u"minute", [1000, 2000, 3000]);

julia> JuChrom.rt2ri(ld)(1.5u"minute") ≈ 1266.712648556876
true

julia> JuChrom.rt2ri(ld)(1u"minute") ≈ 821.4487832484438
true

julia> retentionindex(ld, 1u"minute") ≈ 821.44878324844389
[ Info: extrapolated value
true

JuChrom.InputOutput.FileExistsError — Type

JuChrom.InputOutput.FileExistsError([msg::AbstractString]) <: Exception

File already exists. msg is an optional descriptive error message.

See also IOError.

JuChrom.InputOutput.IOError — Type

JuChrom.InputOutput.IOError([msg::AbstractString]) <: Exception

There was a problem reading or writing a file. msg is an optional descriptive error message.

See also FileExistsError.

JuChrom.InputOutput.Source — Type

JuChrom.InputOutput.Source

Supertype for all data sources.

See also File, Path.

JuChrom.InputOutput.File — Type

JuChrom.InputOutput.File() <: Source

Type that indicates a source is a file.

See also Source, Path.

JuChrom.InputOutput.Path — Type

JuChrom.InputOutput.Path() <: Source

Type that indicates a source is a path.

See also Source, File.

JuChrom.InputOutput.buildxic — Function

JuChrom.InputOutput.buildxic(pointcounts::AbstractVector{<:Integer}, 
ionvec::AbstractVector{<:Real}, intsvec::AbstractVector{<:Real})

Return a sorted list of ions and an intensity matrix with the number of columns equal to the number of ions in the sorted list and the number of rows equal to the number of scans. This is based on a vector that stores the number of ion-intensity pairs per scan (pointcounts), a vector of continuously recorded ions throughout the run (ionvec), and a vector of intensity values associated with those ions (intsvec).

Examples

julia> pointcounts = [2, 3, 2];

julia> ionvec = [85.1, 100.2, 85.2, 99.9, 112.1, 84.9, 100.6];

julia> intsvec = Int64[12, 234, 23, 324, 45422, 21, 523];

julia> mzs, xic = JuChrom.InputOutput.buildxic(pointcounts, ionvec, intsvec);

julia> mzs
7-element Vector{Float64}:
  84.9
  85.1
  85.2
  99.9
 100.2
 100.6
 112.1

julia> xic
3×7 Matrix{Int64}:
  0  12   0    0  234    0      0
  0   0  23  324    0    0  45422
 21   0   0    0    0  523      0

JuChrom.LocalMaxima — Type

JuChrom.LocalMaxima(values::AbstractVector{<:Real}; 
startindex::Integer=firstindex(values), stopindex::Integer=lastindex(values))

Return an iterator that yields the index range of each local maximum until no more local maxima are found. If the maximum is a single peak, the range will have a length of 1; if a plateau maximum is found, the range will have a length greater than 1. The optional keyword arguments startindex and stopindex allow you to restrict the search range for the maximum. Note that startindex and stopindex must be at least two indices apart (e.g., startindex=1, stopindex=3).

See also nextlocalmaximum.

Examples

julia> values = [4, 3, 5, 3, 6, 6, 6, 4, 7];

julia> for lm in JuChrom.LocalMaxima(values); println(lm); end
3:3
5:7

julia> for lm in JuChrom.LocalMaxima(values, startindex=3); println(lm); end
5:7

julia> for lm in JuChrom.LocalMaxima(values, stopindex=7); println(lm); end
3:3

julia> for lm in JuChrom.LocalMaxima(values, startindex=4, stopindex=8); println(lm); end
5:7

julia> for lm in JuChrom.LocalMaxima(values, startindex=4, stopindex=5); println(lm); end
ERROR: ArgumentError: stopindex not greater than startindex + 1: 4 5
[...]

julia> length(values)
9

julia> for lm in JuChrom.LocalMaxima(values, startindex=8); println(lm); end
ERROR: ArgumentError: stopindex not greater than startindex + 1: 8 9
[...]

JuChrom.nextlocalmaximum — Function

JuChrom.nextlocalmaximum(values::AbstractVector{<:Real}; 
startindex::Integer=firstindex(values), stopindex::Integer=lastindex(values))

Return the index range of the next local maximum. If the maximum is a single peak, the range will have a length of 1; if a plateau maximum is found, the range will have a length greater than 1. If no maximum is found, nothing is returned. The optional keyword arguments startindex and startindex allow you to restrict the search range for the maximum. Note that startindex and startindex must be at least two indices apart (e.g., startindex=1, stopindex=3).

See also LocalMaxima.

Examples

julia> values = [4, 3, 5, 3, 6, 6, 6, 4, 7];

julia> JuChrom.nextlocalmaximum(values)  # finds 5
3:3

julia> JuChrom.nextlocalmaximum(values, startindex=4)  # finds pleateau maximum 6, 6, 6
5:7

julia> JuChrom.nextlocalmaximum(values, startindex=4, stopindex=7)  # finds nothing

julia> length(values)  
9

julia> JuChrom.nextlocalmaximum(values, startindex=8)  # implicit stopindex=9
ERROR: ArgumentError: stopindex not greater than startindex + 1: 8 9
[...]

julia> JuChrom.nextlocalmaximum(values, startindex=1, stopindex=2)
ERROR: ArgumentError: stopindex not greater than startindex + 1: 1 2
[...]