Mass spectra

AbstractMassSpectrum

JuChrom.AbstractMassSpectrum — Type

AbstractMassSpectrum{M, I} <: Any

Abstract supertype for mass spectrum containers holding paired m/z values and intensity values.

M is the m/z unit type (Unitful.Units subtype or Nothing), and I is the intensity unit type (Unitful.Units subtype or Nothing).

Concrete subtypes are expected to define mzvalues::AbstractVector{<:Real}, mzunit::Union{Unitful.Units, Nothing}, intensities::AbstractVector{<:Real}, intensityunit::Union{Unitful.Units, Nothing}, and attrs::NamedTuple.

See also MassSpectrum, attrs, intensities, intensityunit, mzcount, mzunit, mzvalues.

JuChrom.attrs — Method

attrs(ms::AbstractMassSpectrum)

Return the metadata NamedTuple associated with a mass spectrum.

See also MassSpectrum, mzvalues, intensities.

Examples

julia> ms = MassSpectrum([100.0, 150.0], [10.0, 20.0]; attrs=(sample="QC",));

julia> attrs(ms).sample == "QC"
true

JuChrom.intensityunit — Method

intensityunit(ms::AbstractMassSpectrum)

Return the intensity unit for a mass spectrum, or nothing if unspecified.

See also MassSpectrum, intensities.

Examples

julia> ms = MassSpectrum([100.0, 150.0], [10.0, 20.0]);

julia> intensityunit(ms) === nothing
true

JuChrom.intensities — Method

intensities(ms::AbstractMassSpectrum; unit=nothing)

Return the intensity values from a mass spectrum.

If unit is specified, intensities are converted with Unitful.uconvert. If the spectrum is unitless and a unit is requested, an error is thrown.

See also MassSpectrum, intensityunit, mzcount, mzvalues.

Examples

julia> ms = MassSpectrum([100.0, 150.0], [10.0, 20.0]u"pA");

julia> intensities(ms) == [10.0, 20.0]u"pA"
true

julia> intensities(ms, unit=u"nA") == [0.01, 0.02]u"nA"
true

JuChrom.mzcount — Method

mzcount(ms::AbstractMassSpectrum) -> Integer

Return the number of m/z points in the spectrum.

See also MassSpectrum, mzvalues, intensities.

Examples

julia> ms = MassSpectrum([100.0, 150.0, 200.0], [10.0, 20.0, 5.0]);

julia> mzcount(ms)
3

JuChrom.mzunit — Method

mzunit(ms::AbstractMassSpectrum)

Return the m/z unit for a mass spectrum, or nothing if unspecified.

See also MassSpectrum, mzvalues.

Examples

julia> ms = MassSpectrum([100.0, 150.0], [10.0, 20.0]);

julia> mzunit(ms) === nothing
true

JuChrom.mzvalues — Method

mzvalues(ms::AbstractMassSpectrum; unit=nothing)

Return the m/z values from a mass spectrum.

If unit is specified, values are converted with Unitful.uconvert. If the spectrum is unitless and a unit is requested, an error is thrown.

See also MassSpectrum, mzunit, intensities, mzcount.

Examples

julia> ms = MassSpectrum([100.0, 150.0]u"Th", [10.0, 20.0]);

julia> mzvalues(ms) == [100.0, 150.0]u"Th"
true

julia> mzvalues(ms, unit=u"kTh") == [0.1, 0.15]u"kTh"
true

MassSpectrum

JuChrom.MassSpectrum — Type

MassSpectrum{T1, T2, T3, T4, T5} <: AbstractMassSpectrum{T2, T4}

Concrete container for a single mass spectrum.

T2 and T4 encode the m/z and intensity unit types (each a Unitful.Units subtype or Nothing). The remaining type parameters capture the stored numeric vectors and metadata.

Fields include mzvalues::AbstractVector{<:Real}, mzunit::Union{Unitful.Units, Nothing}, intensities::AbstractVector{<:Real}, intensityunit::Union{Unitful.Units, Nothing}, and attrs::NamedTuple.

The constructor validates that m/z values are strictly increasing, positive, finite, and that intensity values are finite with matching length.

See also AbstractMassSpectrum, MassSpectrum, attrs, intensities, intensityunit, mzcount, mzunit, mzvalues.