JuChrom.jl

Documentation for JuChrom.jl

JuChrom.AbstractChromatogram — Type

AbstractChromatogram

Supertype of all chromatogram implementations. All subtypes (e.g., FID, GCMS, TIC) contain scantimes, intensities, and metadata.

See also AbstractGC, AbstractGCMS, AbstractFID, AbstractTIC, FID, GCMS, TIC, intensities, scantimes, metadata.

JuChrom.AbstractFID — Type

AbstractFID <: AbstractGC

Supertype of all flame ionization detector chromatogram implementations in JuMS (e.g., FID).

See also AbstractChromatogram, AbstractGC, FID, intensities, scantimes, metadata.

JuChrom.AbstractGC — Type

AbstractGC <: AbstractChromatogram

Supertype of all chromatogram implementations that have no mass-charge ratio (m/z) data (= ions) and therefore have a single intensity value associated with a given scantime (e.g., FID, TIC). The intensities are stored in a vector whose index corresponds to the index of the associated scantime.

See also AbstractChromatogram, AbstractFID, AbstractTIC, FID, TIC, intensities, scantimes, metadata.

JuChrom.AbstractGCMS — Type

AbstractGCMS <: AbstractChromatogram

Supertype of all chromatogram implementations that contain mass-charge ratio (m/z) data (= ions) and associated abundance values (= intensities) and thus can have one or more ion intensity values associated with a given scan time (e.g., GCMS). The intensities are stored in a matrix where the row index corresponds to that of the associated scantime and where the column index corresponds to that of the associated ion.

See also AbstractChromatogram, GCMS, intensities, ions, scantimes, metadata.

JuChrom.AbstractTIC — Type

AbstractTIC <: AbstractGC

Supertype of all total ion chromatogram implementations (e.g., TIC).

See also AbstractChromatogram, AbstractGC, TIC, intensities, scantimes, metadata.

JuChrom.FID — Method

FID(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any})

Construct a FID object consisting of scantimes, intensities, and metadata. Note that the scantimes must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractFID, scantimes, intensities, metadata.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the FID object preserves the types.

Example

julia> FID(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
FID {scantimes: Int64, intensities: Int32}
3 scans; time range: 1 s - 3 s
intensity range: 1 - 956
metadata: 

julia> FID(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict(:name => "sample"))
FID {scantimes: Int32, intensities: Float64}
3 scans; time range: 1 s - 3 s
intensity range: 1.0 - 956.0
metadata: :name

julia> FID([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> FID([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain values less than zero
[...]

JuChrom.GCMS — Method

GCMS(scantimes::AbstractVector{<:Unitful.Time}, ions::AbstractVector{<:Real}, 
intensities::AbstractMatrix{<:Real}; metadata::Dict=Dict{Any, Any}())

Construct a GCMS object consisting of scantimes, ions, intensities, and metadata. Note that the scantimes and the ions must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGCMS, intensities, ions, scantimes, metadata.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the GCMS object preserves the types.

Example

julia> GCMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1])
GCMS {scantimes: Int32, ions: Int64, intensities: Int32}
3 scans; time range: 1 s - 3 s
2 ions; range: m/z 85 - 100
intensity range: 0 - 956
metadata: 

julia> GCMS([1.1f0, 2.1f0]u"s", [35.1f0, 76.2f0], Int64[0 12; 34 956], Dict(:id => 1))
GCMS {scantimes: Float32, ions: Float32, intensities: Int64}
2 scans; time range: 1.1f0 s - 2.1f0 s
2 ions; range: m/z 35.1 - 76.2
intensity range: 0 - 956
metadata: :id

julia> GCMS([2, 1, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [100, 85], [0 12; 34 956; 23 1])
ERROR: ArgumentError: ions not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [85, 100], [0 -12; 34 956; 23 1])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.TIC — Method

TIC(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any}())

Construct a TIC object consisting of scantimes, intensities, and metadata. Note that the scantimes must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractTIC, scantimes, intensities, metadata.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the TIC object preserves the types.

Example

julia> TIC(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
TIC {scantimes: Int64, intensities: Int32}
3 scans; time range: 1 s - 3 s
intensity range: 1 - 956
metadata: 

julia> TIC(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict(:name => "sample"))
TIC {scantimes: Int32, intensities: Float64}
3 scans; time range: 1 s - 3 s
intensity range: 1.0 - 956.0
metadata: :name

julia> TIC([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> TIC([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.intensities — Method

intensities(chrom::AbstractChromatogram)

Return the intensities.

See also AbstractChromatogram, FID, GCMS, TIC, minintensity, maxintensity.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], Int64[0 12; 34 956; 23 1]);

julia> intensities(gcms)
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

julia> fid = FID([1, 2, 3]u"s", Float32[12.0, 956.0, 23.0]);

julia> intensities(fid)
3-element Vector{Float32}:
  12.0
 956.0
  23.0

JuChrom.ioncount — Method

ioncount(gcms::AbstractGCMS) -> Int

Return the number of ions.

See also AbstractGCMS, GCMS, ions.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> ioncount(gcms)
2

JuChrom.ions — Method

ions(gcms::AbstractGCMS)

Return the ions.

See also AbstractGCMS, GCMS, minion, maxion, ioncount.

In the following example, the number type of ions passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> ions(gcms)
2-element Vector{Int64}:
  85
 100

JuChrom.maxintensity — Method

maxintensity(chrom::AbstractChromatogram)

Return the maximum intensity.

See also AbstractChromatogram, FID, GCMS, TIC, minintensity, intensities.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], Int64[0 12; 34 956; 23 1]);

julia> maxintensity(gcms)
956

JuChrom.maxion — Method

maxion(gcms::AbstractGCMS)

Return the largest ion.

See also AbstractGCMS, GCMS, minion, ions, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> maxion(gcms)
100

JuChrom.maxscantime — Method

maxscantime(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool)

Return the time of the last scan. The optional keyword argument timeunit allows you to change the unit of the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, FID, GCMS, TIC, minscantime, scantimes, scancount.

Example

julia> gcms = GCMS(Float64[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxscantime(gcms)
3.0 s

julia> maxscantime(gcms, timeunit=u"minute")
0.05 minute

julia> maxscantime(gcms, timeunit=u"minute", ustripped=true)
0.05

JuChrom.metadata — Method

metadata(chrom::AbstractChromatogram) -> Dict{Any, Any}

Return the metadata.

See also AbstractChromatogram, FID, GCMS, TIC.

Example

julia> gcms₁ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956], Dict(:id => 1))
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; time range: 1 s - 2 s
2 ions; range: m/z 85 - 100
intensity range: 0 - 956
metadata: :id

julia> metadata(gcms₁)
Dict{Any, Any} with 1 entry:
  :id => 1

julia> gcms₂ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956])
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; time range: 1 s - 2 s
2 ions; range: m/z 85 - 100
intensity range: 0 - 956
metadata:

julia> metadata(gcms₂)
Dict{Any, Any}()

JuChrom.minintensity — Method

minintensity(chrom::AbstractChromatogram)

Return the minimum intensity.

See also AbstractChromatogram, FID, GCMS, TIC, maxintensity, intensities.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], Int64[0 12; 34 956; 23 1]);

julia> minintensity(gcms)
0

JuChrom.minion — Method

minion(gcms::AbstractGCMS)

Return the smallest ion.

See also AbstractGCMS, GCMS, ions, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> minion(gcms)
85

JuChrom.minscantime — Method

minscantime(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the time of the first scan. The optional keyword argument timeunit allows you to change the unit of the returned scan time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, FID, GCMS, TIC, maxscantime, scantimes, scancount.

Example

julia> gcms = GCMS(Float64[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minscantime(gcms)
1.0 s

julia> minscantime(gcms, timeunit=u"minute")
0.016666666666666666 minute

julia> minscantime(gcms, timeunit=u"minute", ustripped=true)
0.016666666666666666

JuChrom.scancount — Method

scancount(chrom::AbstractChromatogram) -> Int

Return the number of scans.

See also AbstractChromatogram, FID, GCMS, TIC, scantimes.

Example

julia> fid = FID([1, 2, 3]u"s", [12, 956, 23]);

julia> scancount(fid)
3

JuChrom.scantimes — Method

scantimes(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the scantimes. The optional keyword argument timeunit allows you to change the unit of the returned scantimes. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned values.

See also AbstractChromatogram, FID, GCMS, TIC, minscantime, maxscantime.

In the following example, the element type of the scantime vector passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS(Float32[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantimes(gcms)
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0f0 s
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, timeunit=u"minute")
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.016666668f0 minute
 0.033333335f0 minute
 0.050000004f0 minute

julia> scantimes(gcms, timeunit=u"minute", ustripped=true)
3-element Vector{Float32}:
 0.016666668
 0.033333335
 0.050000004