Basics

JuChrom.AbstractChromatogram — Type

AbstractChromatogram

Supertype of all chromatogram implementations. All subtypes (e.g., FID, GCMS, TIC) contain scantimes, intensities, and metadata.

See also AbstractGC, AbstractFID, AbstractGCMS, AbstractTIC, FID, GCMS, TIC, scantimes, intensities, metadata.

JuChrom.AbstractFID — Type

AbstractFID <: AbstractGC

Supertype of all flame ionization detector chromatogram implementations (e.g., FID).

See also AbstractChromatogram, AbstractFID, AbstractGC, FID, scantimes, intensities, metadata.

JuChrom.AbstractGC — Type

AbstractGC <: AbstractChromatogram

Supertype of all chromatogram implementations that have no mass-charge ratio (m/z) data (= ions) and therefore have a single intensity value associated with a given scantime (e.g., FID, TIC). The intensities are stored in a vector whose index corresponds to the index of the associated scantime.

See also AbstractChromatogram, AbstractFID, AbstractTIC, FID, TIC, scantimes, intensities, metadata.

JuChrom.AbstractGCMS — Type

AbstractGCMS <: AbstractChromatogram

Supertype of all chromatogram implementations that contain mass-charge ratio (m/z) data (= ions) and associated abundance values (= intensities) and thus can have one or more ion intensity values associated with a given scan time (e.g., GCMS). The intensities are stored in a matrix where the row index corresponds to that of the associated scantime and where the column index corresponds to that of the associated ion.

See also AbstractChromatogram, AbstractGCMS, GCMS, scantimes, ions, intensities, metadata.

JuChrom.AbstractTIC — Type

AbstractTIC <: AbstractGC

Supertype of all total ion chromatogram implementations (e.g., TIC).

See also AbstractChromatogram, AbstractGC, AbstractTIC, TIC, scantimes, intensities, metadata.

JuChrom.FID — Type

FID(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any})

Construct a FID object consisting of scantimes, intensities, and metadata. Note that the scantimes must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractFID, scantimes, intensities, metadata.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the FID object preserves the types.

Example

julia> FID(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
FID {scantimes: Int64, intensities: Int32}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 1 - 956
metadata: 0 entries

julia> FID(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict("name" => "sample"))
FID {scantimes: Int32, intensities: Float64}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 1.0 - 956.0
metadata: 1 entry

julia> FID([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> FID([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain values less than zero
[...]

JuChrom.GCMS — Type

GCMS(scantimes::AbstractVector{<:Unitful.Time}, ions::AbstractVector{<:Real}, 
intensities::AbstractMatrix{<:Real}; metadata::Dict=Dict{Any, Any}())

Construct a GCMS object consisting of scantimes, ions, intensities, and metadata. Note that the scantimes and the ions must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGCMS, scantimes, ions, intensities, metadata, totalionchromatogram.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the GCMS object preserves the types.

Example

julia> GCMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1])
GCMS {scantimes: Int32, ions: Int64, intensities: Int32}
3 scans; scantimes: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> GCMS([1.1f0, 2.1f0]u"s", [35.1f0, 76.2f0], Int64[0 12; 34 956], Dict(:id => 4))
GCMS {scantimes: Float32, ions: Float32, intensities: Int64}
2 scans; scantimes: 1.1f0 s, 2.1f0 s
2 ions: m/z 35.1, 76.2
intensity range: 0 - 956
metadata: 1 entry

julia> GCMS([2, 1, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [100, 85], [0 12; 34 956; 23 1])
ERROR: ArgumentError: ions not in ascending order
[...]

julia> GCMS([1, 2, 3]u"s", [85, 100], [0 -12; 34 956; 23 1])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.IonScanOrder — Type

IonScanOrder

Supertype of all ion scan order implementations.

See also LinearAscending, LinearDescending, ionscantimeshift, ionscantime.

JuChrom.LinearAscending — Type

LinearAscending(; start::Real=0, stop::Real=1) <: IonScanOrder

Construct a LinearAscending ion scan order object. It specifies that the ions were scanned in linear ascending order (i.e., smallest ion first, last ion last) during each scan. The time to scan each ion is assumed to be equal and is the result of dividing the total scan interval time equally among the ions. The optional start and stop parameters allow you to limit the interval time during which the ions were scanned in each scan. They specify relative points in the scan interval: 0 ≤ start < stop ≤ 1. The default values are start=0 and stop=1, which means that the scan of the smallest ion started at the beginning of the scan interval and the scan of the largest ion ended at the end of the scan interval. In contrast, setting the start value to 0.5 would indicate that the ions were only scanned during the second half of the scan interval (e.g., because the instrument switched between SIM mode and Scan mode during each scan interval and operated only in the second half of the scan interval in Scan mode).

See also AbstractGCMS, GCMS, LinearDescending, ionscantimeshift, ionscantime, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Example

julia> LinearAscending()
LinearAscending{Int64, Int64}(0, 1)

julia> LinearAscending(start=0.5)
LinearAscending{Float64, Int64}(0.5, 1)

julia> LinearAscending(start=0.1, stop=0.5)
LinearAscending{Float64, Float64}(0.1, 0.5)

julia> LinearAscending(start=0.5, stop=0.5)
ERROR: ArgumentError: start=0.5 and stop=0.5 do not satisfy condition 0 ≤ start < stop ≤ 1
[...]

JuChrom.LinearDescending — Type

LinearDescending(; start::Real=0, stop::Real=1) <: IonScanOrder

Construct a LinearDescending ion scan order object. It specifies that the ions were scanned in linear descending order (i.e., largest ion first, smallest ion last) during each scan. The time to scan each ion is assumed to be equal and is the result of dividing the total scan interval time equally among the ions. The optional start and stop parameters allow you to limit the interval time during which the ions were scanned in each scan. They specify relative points in the scan interval: 0 ≤ start < stop ≤ 1. The default values are start=0 and stop=1, which means that the scan of the largest ion started at the beginning of the scan interval and the scan of the smallest ion ended at the end of the scan interval. In contrast, setting the start value to 0.5 would indicate that the ions were only scanned during the second half of the scan interval (e.g., because the instrument switched between SIM mode and Scan mode during each scan interval and operated only in the second half of the scan interval in Scan mode).

See also AbstractGCMS, GCMS, LinearAscending, ionscantimeshift, ionscantime, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Example

julia> LinearDescending()
LinearDescending{Int64, Int64}(0, 1)

julia> LinearDescending(start=0.5)
LinearDescending{Float64, Int64}(0.5, 1)

julia> LinearDescending(start=0.1, stop=0.5)
LinearDescending{Float64, Float64}(0.1, 0.5)

julia> LinearDescending(start=0.5, stop=0.5)
ERROR: ArgumentError: start=0.5 and stop=0.5 do not satisfy condition 0 ≤ start < stop ≤ 1
[...]

JuChrom.TIC — Type

TIC(scantimes::AbstractVector{<:Unitful.Time}, intensities::AbstractVector{<:Real}, 
metadata::Dict=Dict{Any, Any}())

Construct a TIC object consisting of scantimes, intensities, and metadata. Note that the scantimes must be in ascending order and the intensities must not contain values less than zero.

See also AbstractChromatogram, AbstractGC, AbstractTIC, scantimes, intensities, metadata, totalionchromatogram.

In the following examples, the number types of the arrays passed to the object constructor are explicitly annotated to illustrate that the TIC object preserves the types.

Example

julia> TIC(Int64[1, 2, 3]u"s", Int32[12, 956, 1])
TIC {scantimes: Int64, intensities: Int32}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 1 - 956
metadata: 0 entries

julia> TIC(Int32[1, 2, 3]u"s", Float64[12.0, 956.0, 1.0], Dict("name" => "sample"))
TIC {scantimes: Int32, intensities: Float64}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 1.0 - 956.0
metadata: 1 entry

julia> TIC([2, 1, 3]u"s", [12.0, 956.0, 1.0])
ERROR: ArgumentError: scantimes not in ascending order
[...]

julia> TIC([1, 2, 3]u"s", [-12.0, 956.0, 1.0])
ERROR: ArgumentError: intensity values contain at least one value less than zero
[...]

JuChrom.binions — Function

binions(gcms::AbstractGCMS; ionbin::Function=integerion)

Return a GCMS object in which the ions are binned according to the ionbin function (default function is integer) and the intensities of the binned ions are summed.

See also AbstractGCMS, GCMS, integer, intensities, ions, ioncount.

Example

julia> gcms = GCMS((1:3)u"s", [84.8, 85.2, 100.9], [0 24 12; 0 0 956; 23 0 1])
GCMS {scantimes: Int64, ions: Float64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
3 ions: m/z 84.8, 85.2, 100.9
intensity range: 0 - 956
metadata: 0 entries

julia> intensities(gcms)
3×3 Matrix{Int64}:
  0  24   12
  0   0  956
 23   0    1

julia> gcmsᵢ = binions(gcms)
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
2 ions: m/z 85, 101
intensity range: 0 - 956
metadata: 0 entries

julia> ions(gcmsᵢ)
2-element Vector{Int64}:
  85
 101

julia> intensities(gcmsᵢ)
3×2 Matrix{Int64}:
 24   12
  0  956
 23    1

julia> custom_ionbin(ion) = integer(ion, start=0.9);

julia> gcmsᵢ = binions(gcms, ionbin=custom_ionbin)
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
3 ions: m/z 84, 85, 101
intensity range: 0 - 956
metadata: 0 entries

julia> ions(gcmsᵢ)
3-element Vector{Int64}:
  84
  85
 101

JuChrom.cosine — Function

cosine(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Return the angle between two non-zero vectors, which can be considered a measure of the similarity (i.e., cosine similarity) between the two vectors.

Example

julia> cosine([100, 500, 250], [200, 1000, 0])
0.8978872704229618

julia> cosine([100, 0, 50], [0, 20, 0])
0.0

julia> cosine([100, 500, 250], [10, 50, 25])
1.0

JuChrom.integer — Function

integer(value:::Real; start::Real=0.7) -> Int

Return the integer for the given value that satisfies the following condition: integer - 1 + start ≤ value < integer + start, where 0 ≤ start < 1.

See also AbstractGCMS, GCMS, binions, ions.

Example

julia> integer(29.7)
30

julia> integer(30.0)
30

julia> integer(30.69)
30

julia> integer(29.7, start=0.8)
29

JuChrom.intensities — Method

intensities(chrom::AbstractGC; scanindexrange::OrdinalRange{T, S}) where {T<:Integer, 
S<:Integer}

Return the intensities. The optional keyword argument scanindexrange allows you to select a subset of scans for which intensities are to be returned. Note that the function returns a reference to the intensity vector or a view into the intensity vector, depending on whether the keyword argument selects a subset of scans.

See also AbstractGC, FID, TIC, intensity, minintensity, maxintensity.

Example

julia> tic = TIC([1, 2, 3]u"s", [123, 224, 103])
TIC {scantimes: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensities(tic)  # reference to the data structure
3-element Vector{Int64}:
 123
 224
 103

julia> intensities(tic)[:]  # independent copy of the values
3-element Vector{Int64}:
 123
 224
 103

julia> intensities(tic; scanindexrange=2:3)  # view into the data structure
2-element view(::Vector{Int64}, 2:3) with eltype Int64:
 224
 103

julia> intensities(tic; scanindexrange=2:3)[:]  # independent copy of the values
2-element Vector{Int64}:
 224
 103

JuChrom.intensities — Method

intensities(gcms::AbstractGCMS; scanindexrange::OrdinalRange{T1, S1}, 
ionindexrange::OrdinalRange{T2, S2}) where {T1<:Integer, S1<:Integer, T2<:Integer, 
S2<:Integer}

Return the intensities. The optional keyword arguments scanindexrange and ionindexrange allow you to select specific parts of the intensity matrix to be returned. Note that the function returns a reference to the matrix or a view into the matrix depending on whether the keyword arguments select subranges of the matrix.

See also AbstractGCMS, intensities, minintensity, maxintensity.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensities(gcms)  # reference to the data structure
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

julia> intensities(gcms)[:, :]  # independent copy of the values
3×2 Matrix{Int64}:
  0   12
 34  956
 23    1

julia> intensities(gcms, ionindexrange=1:1)  # all intensities of ion with index 1
3×1 view(::Matrix{Int64}, 1:3, 1:1) with eltype Int64:
  0
 34
 23

julia> intensities(gcms, ionindexrange=1:1)[:]  # independent copy of the values
3-element Vector{Int64}:
  0
 34
 23

julia> intensities(gcms, scanindexrange=1:1)  # intensities of all ions of scan 1
1×2 view(::Matrix{Int64}, 1:1, 1:2) with eltype Int64:
 0  12

julia> intensities(gcms, scanindexrange=1:1)[:]  # independent copy of the values
2-element Vector{Int64}:
  0
 12

julia> intensities(gcms, scanindexrange=1:2, ionindexrange=1:2)
2×2 view(::Matrix{Int64}, 1:2, 1:2) with eltype Int64:
  0   12
 34  956

julia> intensities(gcms, scanindexrange=1:2, ionindexrange=1:2)[:, :]
2×2 Matrix{Int64}:
  0   12
 34  956

JuChrom.intensity — Method

intensity(chrom::AbstractGC, scanindex::Integer)

Return the intensity of a scan by specifying its scanindex.

See also AbstractGC, intensities, scantimeindex.

Example

julia> tic = TIC([1, 2, 3]u"s", [123, 224, 103])
TIC {scantimes: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensity(tic, 2)
224

JuChrom.intensity — Method

intensity(chrom::AbstractGC, time::Unitful.Time; precisetime::Bool=false)

Return the intensity at a given time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. By default, the intensity associated with the scan whose timestamp is closest to the given time is returned. If there is a tie, the intensity of the scan with the later scan time is used. If the optional parameter precisetime is set to true, the specified time must exist in the vector, otherwise an error is thrown.

See also AbstractGC, intensities, scantimeindex.

Example

julia> tic = TIC([1.0, 2.0, 3.0]u"s", [123, 224, 103])
TIC {scantimes: Float64, intensities: Int64}
3 scans; scantimes: 1.0 s, 2.0 s, 3.0 s
intensity range: 103 - 224
metadata: 0 entries

julia> intensity(tic, 1.5u"s")
224

julia> intensity(tic, 1u"s", precisetime=true)
123

julia> intensity(tic, 1.5u"s", precisetime=true)
ERROR: ArgumentError: scantime 1.5 s does not exist
[...]

JuChrom.intensity — Method

intensity(gcms::AbstractGCMS, scanindex::Integer, ionindex::Integer)

Return the intensity of an ion in a scan given the scanindex of the scan and the ionindex of the ion.

See also AbstractGCMS, intensities, minintensity, maxintensity.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensity(gcms, 2, 1)
34

JuChrom.intensity — Method

intensity(gcms::AbstractGCMS, time::Unitful.Time, ion::Real; precisetime::Bool=false)

Return the intensity of an ion at a given time. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. By default, the intensity associated with the scan whose timestamp is closest to the given time is returned. If there is a tie, the intensity of the scan with the later scan time is used. If the optional parameter precisetime is set to true, the specified time must exist in the vector, otherwise an error is thrown.

See also AbstractGCMS, scantimeindex, scancount, ionindex, ions, ioncount.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Float64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> intensity(gcms, 1.9u"s", 85)
34

julia> intensity(gcms, 2.9u"s", 85, precisetime=true)
ERROR: ArgumentError: scantime 2.9 s does not exist
[...]

julia> intensity(gcms, 3u"s", 85, precisetime=true)
23

JuChrom.ion — Function

ion(gcms::AbstractGCMS, ionindex::Integer)

Return the ion at the specified ionindex.

See also AbstractGCMS, ions, ionindex, minion, maxion, ioncount.

Example

julia> gcms = GCMS((1:3)u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> ion(gcms, 1)
85

julia> ion(gcms, 2)
100

julia> ion(gcms, 3)
ERROR: BoundsError: attempt to access 2-element Vector{Int64} at index [3]
[...]

JuChrom.ioncount — Function

ioncount(gcms::AbstractGCMS) -> Int

Return the number of ions.

See also AbstractGCMS, GCMS, ions, ion, minion, maxion.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> ioncount(gcms)
2

JuChrom.ionindex — Function

ionindex(gcms::AbstractGCMS, ion::Real) -> Int

Return the index of the ion. If the ion does not exist, an error is thrown.

See also AbstractGCMS, ions, ion, minion, maxion, ioncount.

Example

julia> gcms = GCMS((1:3)u"s", [85.2f0, 100.1f0], [0 12; 34 956; 23 1]);

julia> ionindex(gcms, 100.1)
2

julia> ionindex(gcms, 201.1)
ERROR: ArgumentError: ion 201.1 does not exist
[...]

JuChrom.ionscantime — Function

ionscantime(δtᵢ::Function, gcms::AbstractGCMS, scanindex::Integer, ionindex::Integer; 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the time at which an ion was actually scanned, given the scanindex, the ionindex and a function δtᵢ that computes the time difference between the timestamp of a scan and the scantime of the ion from the ionindex. The optional parameter timeunit allows you to change the unit of the returned scantime. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value. Note that the timestamp of a scan is assumed to be the time at which scanning of the ion intensities associated with that scan was complete.

See also AbstractGCMS, GCMS, scantimes, scantime, scantimeindex, ions, ionindex, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Float64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantime(δtᵢ, gcms, 2, 1)
2.0 s

julia> ionscantime(δtᵢ, gcms, 2, 2)
1.5 s

julia> ionscantime(δtᵢ, gcms, 2, 2; timeunit=u"minute")
0.025 minute

julia> ionscantime(δtᵢ, gcms, 2, 2; timeunit=u"minute", ustripped=true)
0.025

JuChrom.ionscantimeindex — Function

ionscantimeindex(δtᵢ::Function, gcms::AbstractGCMS, ionindex::Integer, 
time::Unitful.Time; precisetime::Bool=false) -> Int

Return the index of the scan where the scan time for the ion is closest to the specified time, given the ionindex and a function δtᵢ that computes the time difference between the timestamp of a scan and the scantime of the ion from the ionindex. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. If there is a tie, the larger scanindex is returned. If the optional parameter precisetime is set to true, the ion must have been scanned exactly at the specified time, otherwise an error is thrown.

See also AbstractGCMS, GCMS, scantimeindex, ionscantime, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending, scantimes, scantime, ions, ion, ionindex.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1 s, 2 s, 3 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantime(δtᵢ, gcms, 2, 2)
1.5 s

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.5u"s")
2

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.6u"s")
2

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.6u"s", precisetime=true)
ERROR: ArgumentError: ion has not been scanned at the time 1.6 s
[...]

julia> ionscantimeindex(δtᵢ, gcms, 2, 1.5u"s", precisetime=true)
2

julia> ionscantimeindex(δtᵢ, gcms, 1, 1.4u"s")
1

julia> ionscantime(δtᵢ, gcms, 1, 1)
1.0 s

JuChrom.ionscantimes — Function

ionscantimes(δtᵢ::Function, gcms::AbstractGCMS, ionindex::Integer; 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the times at which an ion was actually scanned, given the ionindex and a function δtᵢ that computes the time difference between the timestamp of a scan and the scantime of the ion from the ionindex. The optional parameter timeunit allows you to change the unit of the returned scantimes. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value. Note that the timestamp of a scan is assumed to be the time at which scanning of the ion intensities associated with that scan was complete.

See also AbstractGCMS, GCMS, ionscantime, ionscantimeshift, IonScanOrder, LinearAscending, LinearDescending, scantimes, scantimeindex, ions, ionindex.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Float64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> ionscantimes(δtᵢ, gcms, 1)
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0 s
 2.0 s
 3.0 s

julia> ionscantimes(δtᵢ, gcms, 2)
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 0.5 s
 1.5 s
 2.5 s

julia> ionscantimes(δtᵢ, gcms, 2; timeunit=u"minute")
3-element Vector{Quantity{Float64, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.008333333333333333 minute
                0.025 minute
 0.041666666666666664 minute

julia> ionscantimes(δtᵢ, gcms, 2; timeunit=u"minute", ustripped=true)
3-element Vector{Float64}:
 0.008333333333333333
 0.025
 0.041666666666666664

JuChrom.ionscantimeshift — Function

ionscantimeshift(gcms::AbstractGCMS, ionscanorder::IonScanOrder; error::Real=0.001)

Return a function that calculates the time difference between the timestamp of a scan and the time when an ion was actually scanned, given the index of the ion as an argument. The time difference will be a negative value or zero, since the timestamp of a scan is considered to be the time at which the scanning of the last ion was completed. The returned function relies on the assumption that the duration of each scan is the same across the entire run. The optional keyword argument error allows you to specify the maximum fraction of the average scan time that a scanduration defined by the timestamps of two consecutive scans is allowed to deviate from the average scan time duration.

See also AbstractGCMS, GCMS, IonScanOrder, LinearAscending, LinearDescending, ionscantime, ions, minion, maxion, ioncount, scanduration, scantimes, minscantime, maxscantime.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Float64, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.0 s, 2.0 s, 3.0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> δtᵢ = ionscantimeshift(gcms, LinearAscending());

julia> δtᵢ(1)
-0.5 s

julia> δtᵢ(2)
0.0 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending());

julia> δtᵢ(1)
0.0 s

julia> δtᵢ(2)
-0.5 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(start=0.5));

julia> δtᵢ(1)
0.0 s

julia> δtᵢ(2)
-0.25 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(stop=0.5));

julia> δtᵢ(1)
-0.5 s

julia> δtᵢ(2)
-0.75 s

julia> δtᵢ = ionscantimeshift(gcms, LinearDescending(start=0.25, stop=0.75));

julia> δtᵢ(1)
-0.25 s

julia> δtᵢ(2)
-0.5 s

JuChrom.ions — Function

ions(gcms::AbstractGCMS)

Return the ions.

See also AbstractGCMS, GCMS, ion, minion, maxion, ionindex, ioncount.

In the following example, the number type of ions passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS([1, 2, 3]u"s", Int64[85, 100], [0 12; 34 956; 23 1]);

julia> ions(gcms)
2-element Vector{Int64}:
  85
 100

JuChrom.maxintensity — Function

maxintensity(chrom::AbstractChromatogram)

Return the maximum intensity.

See also AbstractChromatogram, AbstractGC, AbstractGCMS, FID, GCMS, TIC, minintensity, intensities, scantimes, scancount, ions, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxintensity(gcms)
956

JuChrom.maxion — Function

maxion(gcms::AbstractGCMS)

Return the largest ion.

See also AbstractGCMS, GCMS, minion, ions, ion, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxion(gcms)
100

JuChrom.maxscantime — Function

maxscantime(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}]; timeunit::Unitful.TimeUnits, ustripped::Bool)

Return the maximum scan time. The optional second position argument allows you to specify the scan range for which the maximum scan time is returned. The optional keyword argument timeunit allows you to change the unit of the returned scantime. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, FID, GCMS, TIC, minscantime, scantimes, scancount.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> maxscantime(gcms)
3.0 s

julia> maxscantime(gcms, timeunit=u"minute")
0.05 minute

julia> maxscantime(gcms, timeunit=u"minute", ustripped=true)
0.05

julia> maxscantime(gcms, 1:2)
2.0 s

julia> maxscantime(gcms, 1:2, timeunit=u"minute", ustripped=true)
0.03333333333333333

JuChrom.metadata — Function

metadata(chrom::AbstractChromatogram) -> Dict{Any, Any}

Return the metadata.

See also AbstractChromatogram, FID, GCMS, TIC.

Example

julia> gcms₁ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956], Dict(:id => 4))
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; scantimes: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 1 entry

julia> metadata(gcms₁)
Dict{Any, Any} with 1 entry:
  :id => 4

julia> gcms₂ = GCMS(Int64[1, 2]u"s", Int64[85, 100], Int64[0 12; 34 956])
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; scantimes: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> metadata(gcms₂)
Dict{Any, Any}()

julia> metadata(gcms₂)["name"] = "sample"
"sample"

julia> metadata(gcms₂)[:id] = 123
123

julia> metadata(gcms₂)
Dict{Any, Any} with 2 entries:
  "name" => "sample"
  :id    => 123

julia> gcms₂
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; scantimes: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 2 entries

julia> delete!(metadata(gcms₂), "name")
Dict{Any, Any} with 1 entry:
  :id => 123

julia> gcms₂
GCMS {scantimes: Int64, ions: Int64, intensities: Int64}
2 scans; scantimes: 1 s, 2 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 1 entry

JuChrom.minintensity — Function

minintensity(chrom::AbstractChromatogram[, greaterthan::Real])

Return the minimum intensity. The optional argument greatherthan allows you to specify a value greater than which the returned minimum must be. If no intensity is found above the specified value, nothing is returned.

See also AbstractChromatogram, AbstractGC, AbstractGCMS, FID, GCMS, TIC, maxintensity, intensities, scantimes, scancount, ions, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minintensity(gcms)
0

julia> minintensity(gcms, 0)
1

julia> minintensity(gcms, 1000)

JuChrom.minion — Function

minion(gcms::AbstractGCMS)

Return the smallest ion.

See also AbstractGCMS, GCMS, maxion, ions, ion, ioncount.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minion(gcms)
85

JuChrom.minscantime — Function

minscantime(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}}]; timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the minimum scan time. The optional second position argument allows you to specify the scan range for which the minimum scan time is returned. The optional keyword argument timeunit allows you to change the unit of the returned scantime. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, FID, GCMS, TIC, maxscantime, scantimes, scancount.

Example

julia> gcms = GCMS([1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> minscantime(gcms)
1.0 s

julia> minscantime(gcms, timeunit=u"minute")
0.016666666666666666 minute

julia> minscantime(gcms, timeunit=u"minute", ustripped=true)
0.016666666666666666

julia> minscantime(gcms, 2:3)
2.0 s

julia> minscantime(gcms, 2:3, timeunit=u"minute", ustripped=true)
0.03333333333333333

JuChrom.runduration — Function

runduration(chrom::AbstractChromatogram; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the runduration. The optional keyword argument timeunit allows you to change the unit of the returned time interval. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, scantimes, minscantime, maxscantime, scancount.

Example

julia> fid = FID([30.1u"minute", 40.8u"minute", 51.5u"minute"], [12, 956, 23])
FID {scantimes: Float64, intensities: Int64}
3 scans; scantimes: 30.1 minute, 40.8 minute, 51.5 minute
intensity range: 12 - 956
metadata: 0 entries

julia> runduration(fid)
21.4 minute

julia> runduration(fid, timeunit=u"s")
1284.0 s

julia> runduration(fid, timeunit=u"s", ustripped=true)
1284.0

JuChrom.scancount — Function

scancount(chrom::AbstractChromatogram) -> Int

Return the number of scans.

See also AbstractChromatogram, FID, GCMS, TIC, scantimes, minscantime, maxscantime.

Example

julia> fid = FID([1, 2, 3]u"s", [12, 956, 23]);

julia> scancount(fid)
3

JuChrom.scanduration — Function

scanduration(chrom::AbstractChromatogram; error::Real=0.001, 
timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Calculate the periodicity with which the scans were recorded over time. The optional keyword argument error allows you to specify the maximum fraction of the average scan time that the time interval defined by the timestamps of two consecutive scans is allowed to deviate from the average scan time duration. The optional keyword argument timeunit allows you to change the unit of the return value. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, scantimes, minscantime, maxscantime, scancount, runduration.

Example

julia> fid = FID([1.0, 2.0, 3.0]u"s", [12, 956, 1]);

julia> scanduration(fid)
1.0 s

julia> scanduration(fid, timeunit=u"minute")
0.016666666666666666 minute

julia> scanduration(fid, timeunit=u"minute", ustripped=true)
0.016666666666666666

julia> scanduration(FID([1.0, 1.99, 3.0]u"s", [12, 956, 1]))
ERROR: ArgumentError: maximum scan duration variation above threshold: 0.010000000000000009 > 0.001
[...]

julia> scanduration(FID([1.0, 1.99, 3.0]u"s", [12, 956, 1]), error=0.02)
1.0 s

JuChrom.scantime — Function

scantime(chrom::AbstractChromatogram, scanindex::Integer; timeunit::Unitful.TimeUnits, 
ustripped::Bool=false)

Return the scan time by specifying the scanindex. The optional parameter timeunit allows you to change the unit of the returned scantime. All time units defined in the package Unitful.jl (e.g., u"s", u"minute" ) are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned value.

See also AbstractChromatogram, FID, GCMS, TIC, scantimes, minscantime, maxscantime, scancount, ionscantime.

Example

julia> gcms = GCMS([1, 2, 3]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantime(gcms, 2)
2 s

julia> scantime(gcms, 2, timeunit=u"minute")
1//30 minute

julia> scantime(gcms, 2, timeunit=u"minute", ustripped=true)
1//30

JuChrom.scantimeindex — Function

scantimeindex(gcms::AbstractChromatogram, time::Unitful.Time; 
precisetime::Bool=false) -> Int

Return the index of the scantime closest to time in the scantimes. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. If there is a tie, the larger scantime is returned. If the optional parameter precisetime is set to true, the specified time must exist in the scantimes, otherwise an error is thrown.

See also AbstractChromatogram, FID, GCMS, TIC, scantimes, scantime, minscantime, maxscantime, scancount, ionindex.

Example

julia> gcms = GCMS([1.1f0, 2.1f0, 3.1f0]u"s", [85, 100], [0 12; 34 956; 23 1])
GCMS {scantimes: Float32, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.1f0 s, 2.1f0 s, 3.1f0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> scantimeindex(gcms, 1.1f0u"s", precisetime=true)
1

julia> scantimeindex(gcms, 2.1u"s", precisetime=true)
2

julia> scantimeindex(gcms, 2.2u"s", precisetime=true)
ERROR: ArgumentError: scantime 2.2 s does not exist
[...]

julia> scantimeindex(gcms, 2.2u"s")
2

JuChrom.scantimes — Function

scantimes(chrom::AbstractChromatogram[, scanindexrange::OrdinalRange{<:Integer, 
<:Integer}]; timeunit::Unitful.TimeUnits, ustripped::Bool=false)

Return the scan times. The optional second position argument allows you to specify a scan index range. The optional keyword argument timeunit allows you to change the unit of the returned scan times. All time units defined in the package Unitful.jl (e.g., u"s", u"minute") are supported. The optional keyword argument ustripped allows you to specify whether the unit is stripped from the returned values. Note: If no time unit conversion takes place and the time unit is not stripped, the function returns a reference to/view into the data structure.

See also AbstractChromatogram, FID, GCMS, TIC, scantime, minscantime, maxscantime, scancount.

In the following example, the element type of the scantimes vector passed to the object constructor is explicitly annotated to illustrate that the GCMS object preserves the type.

Example

julia> gcms = GCMS(Float32[1.0, 2.0, 3.0]u"s", [85, 100], [0 12; 34 956; 23 1]);

julia> scantimes(gcms)
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 1.0f0 s
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, timeunit=u"minute")
3-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(minute,), 𝐓, nothing}}}:
 0.016666668f0 minute
 0.033333335f0 minute
 0.050000004f0 minute

julia> scantimes(gcms, timeunit=u"minute", ustripped=true)
3-element Vector{Float32}:
 0.016666668
 0.033333335
 0.050000004

julia> scantimes(gcms, 2:3)
2-element view(::Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}, 2:3) with eltype Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}:
 2.0f0 s
 3.0f0 s

julia> scantimes(gcms, 2:3, timeunit=u"minute", ustripped=true)
2-element Vector{Float32}:
 0.033333335
 0.050000004

julia> scantimes(gcms, 2:3)[:]  # return an independent copy of the values
2-element Vector{Quantity{Float32, 𝐓, Unitful.FreeUnits{(s,), 𝐓, nothing}}}:
 2.0f0 s
 3.0f0 s

JuChrom.totalionchromatogram — Function

totalionchromatogram(gcms::GCMS)

Compute the total ion chromatrogram.

See also AbstractChromatogram, AbstractGC, GCMS, AbstractTIC, TIC, scantimes, intensities, metadata.

In the following example, the element type of intensities passed to the object constructor is explicitly annotated to show that the TIC object has the same type.

Example

julia> gcms = GCMS((1.1f0:3.1f0)u"s", [85, 100], Int64[0 12; 34 956; 23 1])
GCMS {scantimes: Float32, ions: Int64, intensities: Int64}
3 scans; scantimes: 1.1f0 s, 2.1f0 s, 3.1f0 s
2 ions: m/z 85, 100
intensity range: 0 - 956
metadata: 0 entries

julia> tic = totalionchromatogram(gcms)
TIC {scantimes: Float32, intensities: Int64}
3 scans; scantimes: 1.1f0 s, 2.1f0 s, 3.1f0 s
intensity range: 12 - 990
metadata: 0 entries

julia> intensities(tic)
3-element Vector{Int64}:
  12
 990
  24

julia> gcms = GCMS((1.1:3.1)u"s", [85, 100], Float64[0 12; 34 956; 23 1])
GCMS {scantimes: Float64, ions: Int64, intensities: Float64}
3 scans; scantimes: 1.1 s, 2.1 s, 3.1 s
2 ions: m/z 85, 100
intensity range: 0.0 - 956.0
metadata: 0 entries

julia> tic = totalionchromatogram(gcms)
TIC {scantimes: Float64, intensities: Float64}
3 scans; scantimes: 1.1 s, 2.1 s, 3.1 s
intensity range: 12.0 - 990.0
metadata: 0 entries

julia> intensities(tic)
3-element Vector{Float64}:
  12.0
 990.0
  24.0