Data import

JuChrom.InputOutput.FileFormat — Type

FileFormat

Supertype of all FileFormat implementations.

See also ChemStationMS.

JuChrom.InputOutput.ChemStationMSReaders.ChemStationMS — Type

ChemStationMS(; datafilename::AbstractString="data.ms")

Returns an object representing the ChemStationMS file format. The optional keyword argument, datafilename, allows you to specify a different name for the data file to be read, instead of the default data.ms (e.g., datasim.ms). Upper and lower case are not considered when specifying the data file name. Note that when using the function importdata, the ChemStationMS data reader expects the Agilent .D folder as the source location, with the specified data file located at the top level of the .D folder. However, you can also provide the full pathname of the data file directly when using the function importdata, in which case the datafilename entry in the ChemStationMS object is ignored.

See also FileFormat, importdata.

Example

julia> ChemStationMS()
ChemStationMS{String}("data.ms")

julia> ChemStationMS(datafilename="DATASIM.MS")
ChemStationMS{String}("DATASIM.MS")

julia> dfolder = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D");

julia> gcms = importdata(dfolder, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 0 entries

julia> datafile = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D/data.ms");

julia> gcms = importdata(datafile, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 0 entries

JuChrom.InputOutput.importdata — Function

importdata(source::AbstractString, fileformat::FileFormat)

Read data from the specified source using the appropriate file format reader. The source may be a file or a directory path, depending on the requirements of the chosen file format reader.

Example

julia> dfolder = joinpath(JuChrom.agilent, "C7-C40_ChemStationMS.D");

julia> gcms = importdata(dfolder, ChemStationMS())
GCMS {scan times: Float32, ions: Float32, intensities: Int64}
2405 scans; scan time range: 191941.0f0 ms - 1.899047f6 ms
5176 ions; range: m/z 29.0 - 562.9
intensity range: 0 - 1186816
metadata: 0 entries