Internals

JuChrom.copy_with_eltype — Function

JuChrom.copy_with_eltype(array::AbstractArray, elementtype::Type)

Create a mutable copy of the array with the type of its elements converted to elementtype.

Examples

julia> JuChrom.copy_with_eltype(Int[1, 2, 3, 4, 5, 6], Float64)
6-element Vector{Float64}:
 1.0
 2.0
 3.0
 4.0
 5.0
 6.0

julia> JuChrom.copy_with_eltype(Float64[1, 2, 3, 4, 5, 6], Int32)
6-element Vector{Int32}:
 1
 2
 3
 4
 5
 6

julia> JuChrom.copy_with_eltype(Float64[1.1, 2, 3, 4, 5, 6], Int32)
ERROR: InexactError: Int32(1.1)
[...]

JuChrom.findclosest — Function

JuChrom.findclosest(A::AbstractVector{<:Number}, x::Number) -> Int

Return the index of the number closest to x in a list A of numbers sorted in ascending order. If case of a tie, the index of the larger number is returned.

Examples

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], 0)
3

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], 1.5)
5

julia> JuChrom.findclosest([-2, -1, 0, 1, 2, 3, 4, 5], -1.5)
2

JuChrom.invert — Function

JuChrom.invert(dictionary::Dict)

Return a dictionary where the values from the input dictionary become the keys. Each key in the returned dictionary maps to a list of all original keys from dictionary that were associated with that value.

Example

julia> d = Dict(:a => 1.0, :b => 2.0, :c => 2.0, :d => 1.0)
Dict{Symbol, Float64} with 4 entries:
  :a => 1.0
  :b => 2.0
  :d => 1.0
  :c => 2.0

julia> JuChrom.invert(d)
Dict{Float64, Vector{Symbol}} with 2 entries:
  2.0 => [:b, :c]
  1.0 => [:a, :d]

JuChrom.LocalMaxima — Type

JuChrom.LocalMaxima(values::AbstractVector{<:Real}; 
startindex::Integer=firstindex(values), stopindex::Integer=lastindex(values))

Return an iterator that yields the index range of each local maximum until no more local maxima are found. If the maximum is a single peak, the range will have a length of 1; if a plateau maximum is found, the range will have a length greater than 1. The optional keyword arguments startindex and stopindex allow you to restrict the search range for the maximum. Note that startindex and stopindex must be at least two indices apart (e.g., startindex=1, stopindex=3).

See also nextlocalmaximum.

Examples

julia> values = [4, 3, 5, 3, 6, 6, 6, 4, 7];

julia> for lm in JuChrom.LocalMaxima(values); println(lm); end
3:3
5:7

julia> for lm in JuChrom.LocalMaxima(values, startindex=3); println(lm); end
5:7

julia> for lm in JuChrom.LocalMaxima(values, stopindex=7); println(lm); end
3:3

julia> for lm in JuChrom.LocalMaxima(values, startindex=4, stopindex=8); println(lm); end
5:7

julia> for lm in JuChrom.LocalMaxima(values, startindex=4, stopindex=5); println(lm); end
ERROR: ArgumentError: stopindex not greater than startindex + 1: 4 5
[...]

julia> length(values)
9

julia> for lm in JuChrom.LocalMaxima(values, startindex=8); println(lm); end
ERROR: ArgumentError: stopindex not greater than startindex + 1: 8 9
[...]

JuChrom.name — Function

JuChrom.name(::Type)

Return the name of the type.

Example

julia> gcms = GCMS(Int32[1, 2, 3]u"s", Int64[85, 100], Int32[0 12; 34 956; 23 1]);

julia> JuChrom.name(typeof(gcms))
GCMS

JuChrom.nextlocalmaximum — Function

JuChrom.nextlocalmaximum(values::AbstractVector{<:Real}; 
startindex::Integer=firstindex(values), stopindex::Integer=lastindex(values))

Return the index range of the next local maximum. If the maximum is a single peak, the range will have a length of 1; if a plateau maximum is found, the range will have a length greater than 1. If no maximum is found, nothing is returned. The optional keyword arguments startindex and startindex allow you to restrict the search range for the maximum. Note that startindex and startindex must be at least two indices apart (e.g., startindex=1, stopindex=3).

See also LocalMaxima.

Examples

julia> values = [4, 3, 5, 3, 6, 6, 6, 4, 7];

julia> JuChrom.nextlocalmaximum(values)  # finds 5
3:3

julia> JuChrom.nextlocalmaximum(values, startindex=4)  # finds pleateau maximum 6, 6, 6
5:7

julia> JuChrom.nextlocalmaximum(values, startindex=4, stopindex=7)  # finds nothing

julia> length(values)  
9

julia> JuChrom.nextlocalmaximum(values, startindex=8)  # implicit stopindex=9
ERROR: ArgumentError: stopindex not greater than startindex + 1: 8 9
[...]

julia> JuChrom.nextlocalmaximum(values, startindex=1, stopindex=2)
ERROR: ArgumentError: stopindex not greater than startindex + 1: 1 2
[...]

JuChrom.RetentionIndexStyle — Type

JuChrom.RetentionIndexStyle

Supertype for traits related to the availability of retention index data.

See also HasRetentionIndexData, HasNoRetentionIndexData.

JuChrom.RetentionIndexStyle — Method

JuChrom.RetentionIndexStyle(::Type)

Return the RetentionIndexStyle trait assigned to the specified data type.

See also AbstractChromatogram, HasRetentionIndexData, HasNoRetentionIndexData.

Examples

julia> tic = TIC([1, 2, 3]u"s", [12, 956, 1]);

julia> JuChrom.RetentionIndexStyle(typeof(tic))
JuChrom.HasNoRetentionIndexData()

julia> tic = RiTIC([1, 2, 3]u"s", "Kovats", Int64[100, 200, 300], [12, 956, 1]);

julia> JuChrom.RetentionIndexStyle(typeof(tic))
JuChrom.HasRetentionIndexData()

JuChrom.HasNoRetentionIndexData — Type

JuChrom.HasNoRetentionIndexData <: JuChrom.RetentionIndexStyle

Data type that functions as a trait indicating the absence of retention index data.

See also RetentionIndexStyle, HasNoRetentionIndexData.

JuChrom.HasRetentionIndexData — Type

JuChrom.HasRetentionIndexData <: JuChrom.RetentionIndexStyle

Data type that functions as a trait indicating the availability of retention index data.

See also RetentionIndexStyle, HasNoRetentionIndexData, RiFID, retentionindexname, retentionindices, maxretentionindex, minretentionindex.